[roll] Roll fuchsia [fidl] Server side dynamic test suite
This CL makes a new dynamic test suite focused on server-side tests
only. The idea is that it would be complemented by a separate
client-side test suite, possibly sharing some helper logic, and replace
the existing dynamic test suite. The client-side and server-side test
suites are both big projects and sharing definitions between the two
only reduces a small part of the overall work so it seems better to
solve one first and then move on to the other.

This test suite is meant to be basic and doesn't include some of the
syntactic niceties that exist for the existing dyn_suite. The approach
is to keep things simple and evolve based on needs. Likely some aspects
of the existing dyn_suite will be eventually appear in this suite.

Each binding implementation is entirely detached from the harness
implementation so new out-of-tree bindings can easily be added. The
only piece that is centralized is that all of the build rules are listed
in a shared GN group - but this is merely for convenience.

All existing bindings have test servers implemented, except for Dart.
Dart is not included now because of some crashes that have been hard to
debug, but the intention is to eventually add Dart support.

Original-Bug: 102339
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/694986
Original-Revision: 58c512c93fb912e55cb61bd7eb176d201d343515
GitOrigin-RevId: 00cea0ec5a9d36e8868b75ac0a1cd948c2345a69
Change-Id: Id143e8a51cd855ad4342f8635790d48a661a2e7c