[roll] Roll fuchsia [netstack3] Don't include SLL in RS w/ unspecified src
Previously we incorrectly sent router solicitation messages with the
source link-layer option when the source address is unspecified which
is an RFC violation.

This change introduces an either type for `Serializer`s
(`EitherSerializer`) to the `packet` crate. We cannot impl `Serializer`
for `Either` in the `packet` crate because `Either` implements
`GrowBuffer` and `ShrinkBuffer` and there is a blanket impl of
`Serializer` for all types that implement both `GrowBuffer` and
`ShrinkBuffer`. Also note that `packet`'s `Either` type is documented
to only support buffers.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/694937
Original-Revision: 056a491f0f5801c6e5c745a483f92ca36eebe009
GitOrigin-RevId: a6deb48d1692cd9c6fd6233268725a8b190a6cc1
Change-Id: Ib850e16160e3c10ebc50afbfbf96dc10283ac564