[roll] Roll fuchsia [run-test-suite] Fix multibyte UTF8 bugs
AKA - why not to write your own ANSI filter.
This fixes some issues related to UTF8 characters that are multiple
bytes. The previous iteration of the ANSI filter crashed if a half of
a multibyte UTF8 character was written in one call to write(), and the
second half was written in a subsequent write. This broke some
assumptions around indices.

It also fixes an issue where invalid UTF8 bytes would crash the writer
for similar reasons. In this case the parser produces a replacement
character, which may be a different length from the bytes it replaces.
This again broke assumptions around indices.

This issue was exposed when I tried to move stdout buffering from
before the ANSI filter to after it. Buffering before writing to the
filter probably increased the odds that a full character would always
be present when writing to the filter.

Testing: new unit tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/696451
Original-Revision: 866bf761b277495a89be0a7e0ed4764f74c16c74
GitOrigin-RevId: 0fa6cf96a0e7423d35b8af706aac468f596b8f74
Change-Id: Ibbad92aed7320d753ab1c1f8646b4bbfdaa1d33d