[roll] Roll fuchsia [lowpan_driver_common] Don't remove routes that are still in use.
If there are multiple addresses with the same subnet, then removing
one of them would cause the on-link subnet route to be removed,
making the other devices on the subnet unreachable. This change
adds a table to keep track of addresses being added and removed
from the interface so that the subnet is only removed when all
of the addresses with that prefix have been removed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/689687
Original-Revision: cd65ff05c3fb5bbb6df0436370a772bcdbbc4074
GitOrigin-RevId: 81c363ee0306370d0426b2a47281956008703bca
Change-Id: I5fa0de82ce392e3596acd17b56f60983124b8910