[roll] Roll fuchsia [package-directory] RootDir::new error variant for missing meta.far
RootDir clients may want to know if the error was caused by the
meta.far not being present in storage (e.g. pkg-cache activating cache
packages, in which case an absent cache package is not an error), and
without this new variant it is difficult for clients to know which
variant corresponds to a missing meta.far (and we weren't guaranteeing
it wouldn't change).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/697125
Original-Revision: 9ea8a42c2ae12c4ed4ececc0fd9eff80e87e2d09
GitOrigin-RevId: b7ff1008d86e9a0426b5e4869789bd04b68317e4
Change-Id: Ie9bbe9ccb9a7538bd6159417c32a360f644ce963