[roll] Roll fuchsia [fidldoc] Work on syscall documentation generation.
Adds support to fidldoc to update a directory in-place, leaving
non-generated files untouched. This allows generating syscall
documentation while leaving the hand-written index files in place.

Updates the sycall template to pass the configuration in and hooks up
"generator instructions" in the output files on how to regenerate them.

Adds a script to run fidldoc to generate the syscall documentation.
Since the syscall documentation does not exist in fidl yet, this is done
to the side of the existing update-docs-from-fidl script. When we switch
over the documentation source, the new script will overwrite the old
update-docs-from-fidl.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/696188
Original-Revision: 582f7ba2ce04312d6746ff8ed2091a1cf2ae1bda
GitOrigin-RevId: 48344c21bc96813f51907d41acd121b0210a9257
Change-Id: Ie81d5139ea3137d053854cd5801e55abefeaf8f0