[roll] Roll fuchsia [blobfs][host] Option for exporting compressed blob versions.
Adds an option to provide a path where compressed versions of blobs are
placed, and adds an entry to exported json to provide resulting path
information if the compressed version is generated.

Adds a test to verify that this new method works and generates a valid
blob.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/692602
Original-Revision: 39f0f100d3d144a0aa6945cba7d2a8aed29c0e99
GitOrigin-RevId: c06a4c6146cf6ac7a4d398230f54d0e028f0429d
Change-Id: I0ad207004a5e8815dd29221bcf6ee07d47a16178