[roll] Roll fuchsia [blobfs] Move write buffer to WriteInfo
As all reading is done now using the pager path, there is no need for
the VMO to be accessible outside of the write path.

Remove unused target compression size as the new direction for streaming
writes is to use offline compression.

Remove initializer for write_info_ when opening existing blobs, which
should save a bit of memory (it was only required for the target
compression size).

Test: No functional change.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/695524
Original-Revision: 43f9d27e2fcc3b9606bc72d37728deb8782d64f7
GitOrigin-RevId: d4b4112ff06bb70d036f464518353fa6b788a977
Change-Id: Ia2d4f3e79cb109d017cb5f84f0def3909ff2b855