[roll] Roll fuchsia [cmx] Remove overnetstack-with-serial.cmx
I'd try to migrate it, but it's currently unused, so I don't know how to
test that I successfully migrate it.
The CF team will gladly help recreate this as cml if needed when the
time comes for this to be in use.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/697154
Original-Revision: 623ed2d4e93b248b7ff6b839e01c48b049094f92
GitOrigin-RevId: 193d37c94b4256e4bb9dbb66c3965f8220532d74
Change-Id: I5efa7794022e4a8ce680b70c2ad219c5d1ebe4a5