[roll] Roll fuchsia [bt][fastpair] Evict procedures after default deadline
The GFPS spec states that the shared secret in a Pairing Procedure must
be evicted after 10 seconds (if no update is received).

Implement Future for a Fast Pair Procedure - the Future resolves if the
shared secret deadline is met. Evict the procedure in this case as the
procedure is no longer valid.

Implement Drop for a Fast Pair Procedure - if it's evicted (and
therefore dropped), any outstanding pairing request will be rejected.

Original-Bug: 99757
Test: fx test bt-fastpair-provider-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/693931
Original-Revision: f181f0738c5f203e3980f424a4d2a8c6551cd77c
GitOrigin-RevId: 32693698028f2acf16e65f2987daf88807456583
Change-Id: Id23cca3819aa2edd6a099ad1c2ec1d62a0e09351