[roll] Roll infra/recipes [git] Delete `directory` arguments
Some git module functions accepted a `directory` argument that they
would pass into the top-level `git -C` flag. This was rarely used
because it has the same effect as wrapping the git call with
`api.context(cwd=...)`, which is the preferred generic way of doing this
in recipes. And with only one exception, every time the `directory`
argument was used it was unnecessary because the code was already
wrapped with an `api.context(cwd=...)` context manager that put it in
the correct directory.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/698051
Original-Revision: 29ba328b62a136b962fb35d3c5331231dd68174a
GitOrigin-RevId: 56ace8b9bb3bacf09958b1723b6e6be43d449700
Change-Id: If90b74e5cb81f430b1de701070dba38ad924c62f