[roll] Roll fuchsia [build] Only exclude gnu linker symbols when is_gcc
These symbols are created when DSO's are linked by gcc
toolchains. They show up even when lld links shared
objects against these ld.gold linked shared objects
too. For that reason it wasn't enough to just exclude
these symbols when is_gcc because those libraries
may not be built with gcc but are linking against
a library linked with a gnu linker. Libraries are
linking against the stubs so removing these symbols
from the stubs is enough to ensure lld linked
shared objects do not pull in these symbols so we
can just use is_gcc in the abi check as well. It
is important to note that the abi check is checking
against the actual shared object and not the stub
which is why the is_gcc check must exist in
verify_public_symbols as well as zircon_toolchain.

Fixes: 99723
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/689843
Original-Revision: 4dd58c9c4a408cb5c92efb884a42d16ca2553492
GitOrigin-RevId: 28596c1b374e92aa38032c979ce187c7925de6f1
Change-Id: Ie9d31ee7166d1f606f49b85c285a38eb0ab6f856