[roll] Roll fuchsia [kernel][coresight] Remove tracking of base address from RomTable
This information is not needed, and to store this is  awkward in the
MMIO hwreg context, as the provided hwreg IO provider already
encapsulates the base address. Consequentially, since we no longer have
access to the base address, we make the callback passed to
RomTable::Walk be a function of the offset of each found component.

Moreover, we move `max_offset` from being a parameter to RomTable's
constructor to one of Walk(), as it is not the state of the table itself
but rather of the aperture implicitly being passed to Walk().

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/694687
Original-Revision: aebff48ab8f0aa7bac4f2e1fbb04d1f0fc9020e7
GitOrigin-RevId: f0dd47c675146fde5e52ffd01f4efe6120c7efe5
Change-Id: I05168be122d882494e42a5b29a4ee22d9b41eaed