[roll] Roll fuchsia [governance] Rename "UI" area back to "Graphics".
This is a partial revert of fxrev.dev/662233. That CL renamed
'Graphics' to 'UI' as an area for the Glossary, but also renamed the
functional area for the API Council.

Consolidating the various UI-adjacent areas into a single 'UI' area
might be a reasonable thing to do, but that would require agreement
from the API Council.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/697957
Original-Revision: 4ceb2a9e0197bd0e4e5179cf73a7a9f1e1ef3c99
GitOrigin-RevId: 1d2b13690727f37c8cbd99b5f41d8c02bc35dace
Change-Id: I74a76bdcbba8a4c06e498340b18db642254dfed9