[roll] Roll fuchsia [governance] Make FEC the owners of governance/.
Before this change, `//docs/contribute/governance` was owned by the
docs team, and only the `rfcs/` subdirectory was owned by the
FEC. This CL adds the FEC to the OWNERS file. This probably didn't
matter much in the past, but there are now multiple FEC members who
are not global approvers, so it's worth fixing this.

This also removes some members of the docs team from the OWNERS file,
in the spirit of least privilege. The tech writing leads should make
sure to loop in the FEC for any governance-related changes.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698059
Original-Revision: 8474f5f850ef97953fec78a0dee70e4f8bb895c0
GitOrigin-RevId: 83a90334a383eb8a9e5a7d9e6ed484c26cb73f6b
Change-Id: I2f39e3a83bccd254cedcc4a3db21439a82269fed