[roll] Roll fuchsia [ffx] Refactor emu to separate config from launch
As described in go/ffx-emu-config-design, before we can implement our
plan for configuration editing in the emulator launcher, it's necessary
to separate the process of building a Launch Configuration from using it
to perform the launch.

This CL adds a new function to the EmulatorEngine trait, stage(), which
performs what was the first section of the start() function, i.e.
building the Launch Configuration. The stage() function processes the
configuration, populating the engine's flag data. Rather than return it
directly, we also add a function build_emulator_cmd() to the trait, so
both the engines or the commands themselves can reproduce the Command as
needed. That Command is now passed in to the start() function
explicitly.

This refactor enables us to do see, use, and modify the Launch
Configuration directly from the ffx emu commands themselves, enabling us
to save the Launch Config to the instance directory, edit it on the fly,
and display it to the user on request.

Original-Bug: 98268
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/697003
Original-Revision: a0c40b510832580343fa8c8d6bcc2ca26e95105b
GitOrigin-RevId: 0256ec57009462767e41ccc9de81954e3e940e1c
Change-Id: Ib7e5e4d84e8469fedaa76c5e8d11bcf81ffdffd0