[roll] Roll fuchsia [starnix] Remove redundant is_shutdown test from UnixSocket::read
If this condition is met, the following peek/read call will return MessageReadInfo::default() because the queue is empty, the EAGAIN return would be skipped because the socket is shut down, and MessageReadInfo::default() would be returned anyway. So the check is redundant.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698403
Original-Revision: bc4a83a32c93aa8b17c2fe2f3318ca9996a3120e
GitOrigin-RevId: b1d02744fe8e5ed63317dbb7eaa42b61a561ecf7
Change-Id: I838c23957aaf1ed0672b6841416ab8e18fcb101c