[roll] Roll fuchsia [audio][tools] wav_player resilience to partial frames
If a WAV file ends with an incomplete frame, then for that
final frame wav_player will play the bytes that ARE included,
and pad the rest of the frame with silence. Previously, the
utility sent a packet.payload_size that was not a multiple
of the frame size, which caused the renderer to disconnect.
'bytes_per_packet_' (which determines packet.payload_size)
is defined as a multiple of frame_size_, so this "round up"
will never exceed the bounds of our packet.

This issue was uncovered by a WAV file provided to us from
another team, during hardware validation.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/688721
Original-Revision: 8a0006fbdfed6eab6fc190ab5498814e9a9e122d
GitOrigin-RevId: c69ade9ba2240ba39b42d0021197e2d71d2172b5
Change-Id: I68d1c8f2b6a0c975d8e69933a7c85b97ef0474f0