[roll] Roll fuchsia [virtualization][vsock] Add RX support to the RW state
This CL adds support for guest RX to the ReadWrite state without
actually wiring it in through the connection to be used by the device.

This CL also addresses an overflow bug in the credit system. Credit is
meant to be a running count which (unlike in C++) needs to be explicit
during the overflow in Rust.

Tested:
 - Data integrity tests
 - Overflow tests

Original-Bug: 97355

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/695742
Original-Revision: f49c0b79ae85923c69ea63097c86090c2e6bbaae
GitOrigin-RevId: 9bc81f9554fe0e4aa741099e373389cb4a976653
Change-Id: I7dff447d396c17b49fe4d8b499f4c2d992b2a4b1