[roll] Roll fuchsia [virtualization][vsock] Move socket outside of RefCell
Async read and writes require a mutable reference, but we were never
using the async functionality due to the credit system. This means we
can just drop down the sync socket when reading and writing instead, and
do away with the RefCell entirely.

The key benefit here is that the RX path won't need a write lock on the
state, just a read lock.

Original-Bug: 97355
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698224
Original-Revision: fa57cc603747f9d27e21a393f882c85cbee19eb4
GitOrigin-RevId: 58d7e48c7dca17ec07671b2542f54b3f08773cf4
Change-Id: I5c91dab2f29019703c25c5582f68f068417963f8