[roll] Roll fuchsia [kernel][vm] Clean up PrepareForWriteLocked args
Fix the order of the arguments to be consistent with the rest of the
functions that take a page request, with the offset and length coming
first, followed by the page request. No change in behavior.

Original-Bug: 63989
Run-All-Tests: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698606
Original-Revision: 0a1b778ae357a009acf27e5bb6d2f3eab4b31a1b
GitOrigin-RevId: 755157969ae59ba196546a80d48894a8b1a56db7
Change-Id: I661faf8fcbb3a4b3271ad6ee70edf4096a560490