[roll] Roll fuchsia [debugger] Convert to using DwarfUnit.
Attempting to use the UnitSymbolFactory for the first time revealed that
the offset we actually want is from the DwarfUnit, not the CompileUnit.

This patch adds a way to get the DwarfUnit from the CompileUnit, removes
the now-unnecessary DIE offset on CompileUnit, and adds clarifying
comments around the difference between DwarfUnit and CompileUnit.

Original-Bug: 79529

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698566
Original-Revision: c8d4b98b134e0a6c703eceed52a1515d5d49c90a
GitOrigin-RevId: 8a9489ad4e8ab5ef6448a40027f9ff1f48786895
Change-Id: I77738e0b8d13a5e8e07cfcf795744b35d33c17b9