[roll] Roll fuchsia [fdf][bind] Migrate fidl_offers in props to gened
Migrate the device properties that are added from the fidl offers on a
device to the new generated bind libraries. Previously the property had
a value of true, now it is the enum from the bind library we generate
from the FIDL protocols.

Original-Bug: 100242
Tests: Updated the integration and unit tests and confirming they work
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698624
Original-Revision: f3cdb53170497e035c622bd2e80668aa72344393
GitOrigin-RevId: 3102205340e6c06df83ed64d6fb5f36473173421
Change-Id: Ibccdd6be25e523043daf15fbff7614fcb2103b18