[roll] Roll fuchsia [virtualization] Implement RESOURCE_ATTACH_BACKING
The RESOURCE_ATTACH_BACKING command allows for the guest to specify a
set of guest-physical pages that are associated with a resource. These
guest-physical pages need not be physically contiguous so the backing
memory is provided in the form of a scatter-gather list.

Once the backing memory is associated with a resource, the driver can
then issue TRANSFER_TO_HOST_2D commands to copy data from the backing
memory into the host-private resource (ex: a vmo shared with Flatland).

Original-Bug: 102870
Test: fx test virtio_gpu_unittests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/695690
Original-Revision: b4f1cb6cf504600c21ac2fa5db73744139253282
GitOrigin-RevId: 8d6548f28d00cf12d162b82786052bafe87c985c
Change-Id: If9e04cd50958046625a2cb038d1da96b0387af6a