[roll] Roll fuchsia [archivist][tests] Use the same archivist for testing v1 protocols
Instead of two separate archivist, we can use the same one to test
protocols that are only used in the v1 world.

One log is reduced to "warn" severity, as the "log connector" protocol
is typically not available (provided by appmgr) unless the test itself
is mocking it and providing it. Instead of allowing "ERROR" logs in the
test we can just decrease severity to warn.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698587
Original-Revision: 5d9cc484221d9639b1efb2ef7eba76a56ac08b1e
GitOrigin-RevId: f217d149ed918784a72bc8834d32849f7a6967ec
Change-Id: Ia04df75ee0db65ba8a6b3374c6dbad01d10294de