[roll] Roll fuchsia [settings] Add fidl_storage to factory
Fidl storage factory support is added with this CL. The Fidl storage
factory does not differ much from the current DeviceStorageFactory used
for stash. It only instantiates a different backing storage and has
different serialization requirements for the types it can prepare for
storage.

The /data/storage directory is used to store the data files, and the
storage agent is now able to choose between stash and persistent fidl
storage when writing. The agent keeps track of both factories, and will do
so until all settings have migrated away from stash.

The vast majority of file changes are are updating the storage trait
impls for the controllers to mark which storage is being used to track a
particular setting.

Currently, no settings have been converted to using fidl storage. This
is mostly plumbing in preparation for the first setting to be converted
to persistent fidl.

Original-Bug: 91407
Test: fx test -o setui_service_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/693310
Original-Revision: fd0aa257b1defd183e23833aab652e9668d59a9c
GitOrigin-RevId: b6597654fcbb165cffff8972958a3dd6e5f29924
Change-Id: I4c19ec968ba32dcde36bc7653b6b6218a3737966