[roll] Roll fuchsia [hypervisor] Refactor FPU save and restore
Move FPU save and restore out of the core VCPU loop, so that we can
speed up VM entry and exit.

We only need to save and restore the FPU state when we are returning
back to host user-space, and when migrating CPUs, this is automatically
handled by the context switch. In all other cases, we can avoid the cost
of the FPU save and restore.

Original-Bug: 95763
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/697823
Original-Revision: 3cead0b778eaaf3ff39faace7fb539dbdd9e25b5
GitOrigin-RevId: e71c02fd5523725f34fa2331d7f6e0f0a6fe7d43
Change-Id: Ibc3a3f8e0220b3c1278434b596f9f2f56d63a6b7