[roll] Roll fuchsia [overnet] rewrite stream framer library
This rewrites the stream-framer library from being a complex state machine
into a synchronous helpers that manage serializing and deserializing the
bytes. In addition, it lifts management of the deframing timeout from the
Deframer and over to the single client that uses this functionality. This
allows further simplification of the Deframer.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/699129
Original-Revision: 49d599135591d999a29ea72238d0b9b9dfb01d29
GitOrigin-RevId: b725fd34f0cc68b30461539fea82d40f81e5ce7f
Change-Id: I7f4632346d46fa52770e708cdd864e13b77f053e