[roll] Roll llvm-project/libc [libc] Linux threads - store a ptr to the thread attribs in the start args.
Previosly, a pointer to the thread data structure was stored in the
start args. However, the thread data structure need not have the
lifetime of the thread. On the the other hand, thread attributes are
stored on the thread stack so they live as long as the thread lives.

GitOrigin-RevId: bf79c6a184de79f3977ef34cd4fe56cf4e1f39bf
Original-Revision: a3e69d565d5ae55578b85727ba1f2b09de52d5a8
Change-Id: I37b3647a48d841f3bdb107ee70b9525602c772a1