[roll] Roll fuchsia [starnix] Reclaim shared memory in binder
This change adds logic to reclaim shared memory when it is freed, and
also updates the allocator to use the reclaimed memory.

The new allocator works as follows:

  1. Check if the allocation fits between the current offset and either
     the end of the shared vmo, or the start of the next allocation.
  2. If the allocation does not fit, attempt to "wrap" the allocator
     back to the start of the vmo. If the allocation does not fit at the
     start of the vmo, then return ENOMEM. If it does, allocate at the
     start and update the offset accordingly.

Clients will retry on ENOMEM, and given the pattern of allocations this
allocator should be sufficient for quite some time (since the
SharedMemory instances are shortlived).
.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698788
Original-Revision: 204436179c9234cf5652e26c41a736ba0126a93d
GitOrigin-RevId: 0f221949c7d714ec92a6bede338a310ee18d0a6e
Change-Id: Iddd171334da1431e267bf5e3d05d7c71ce0d11ff