[roll] Roll fuchsia [storage][perf] New filesystem benchmark framework
This framework is intended to replace the existing
start-storage-benchmarks + odu + tracing combo.

The main reason for the new framework is the ability for the benchmarks
to remount the filesystem. Remounting the filesystem should clear any
cached data from the setup phase of the benchmark so we can benchmark
cold reads and writes.

Original-Bug: 94248
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/695563
Original-Revision: 58003d53880c190f72ff2f8fcf1860610cbb13a0
GitOrigin-RevId: aad616710d18d083c3e53eea76e3685830165201
Change-Id: I9db8b0f668d5bfbe9b75e2936d4397da6bd45cad