[roll] Roll fuchsia [build] Don't always build the full SDK
We should build the SDK in SDK builders for instance.
Most often, we build a product and we use some SDK tools for build
actions, but don't need to build the full SDK.

This is important especially because the SDK cross-compiles host tools
with all supported host toolchains, rather than with just the toolchain
for the build host. For instance if building and testing core.x64 there
is no need to build ffx for host_arm64.

This makes some clean builds >20% faster!!
This also significantly cuts down on iteration times when working on ffx
or ffx plugins or their dependencies.

Original-Bug: 101211

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698968
Original-Revision: 6633ffccfbc61c38cd881302aa8f757fc5b43ab6
GitOrigin-RevId: eeeb548d3404194c98054fe5c4fea858ecbffe12
Change-Id: Ib05461243a9f7fffb7b80f71dcc7234b57bf7ffe