[roll] Roll fuchsia [ffx][config] Environments know their own path.
This makes it so that an Environment knows where it came from, so you
don't have to keep track of that separately. From this starting point
we'll be able to build up some additional capabilities in the
environment object as well.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/699075
Original-Revision: f281e5e2ec45345982bd697d44d1249cd125e20a
GitOrigin-RevId: bbe0d271d4446dc1c299a08364e5e81a2f49cb5e
Change-Id: I8d36f1fcaa2b9850f6fd892ae2965e5639d784bb