[roll] Roll fuchsia [ffx][config] Move query parsing out of storage
This moves the query parsing logic completely out of the storage::Config
object, making it a great deal simpler.

As part of that, it moves a bunch of things that were lying all over the
codebase and puts them in more logical places (like get_config from the
api root to query and validate_write_query from Config to query). We get
to only validate the queries in one place as well, which is nice.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/699076
Original-Revision: e5ee97cbb50856a2b54dbbc9bb93b282bba5dfa7
GitOrigin-RevId: ca225947c12de24d51fc9b3c7c4bd86367c522cd
Change-Id: Ia2a6d4c95878c7831e2a47c19ea497f777ccc6c0