[roll] Roll fuchsia [stream][test] Reduce iterations in ReadShrinkRace
Reduce the number of iterations and wait time in ReadShrinkRace, since
it is an inherently slow test.

https://fxrev.dev/698442 initially bumped the number of iterations of a
couple of tests to better test some races. This test is causing some
other builders to time out on tests because of the number of iterations
it is now doing.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/699366
Original-Revision: 06cb2e65b45da20a5acea3a075af4f47ccb8c6d2
GitOrigin-RevId: 320b8fbdf0e9dbb065fa14f1cefe52e982bfc6c3
Change-Id: I445364d13cb373f708dfb0e038e41ecc4dc10acf