[roll] Roll fuchsia [kernel][arm64] arm64 usercopy variant
This introduced a slight rewrite of usercopy.
The primary goal is to enable switching to ldr/stt instructions.

This CL also introduces arm64_user_copy_test GN template, for
adding a test of a variant, based on the loop unrolling that was
originally done for memcpy.

The next steps involve allowing to align either src or dst on the
loop prologue.

The kernel implementation has not changed at this point.

Original-Bug: 59284
Test: fx test -v -o arm64-user-copy-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/693863
Original-Revision: 38e9ee210f9c9243e930f67abe59e85987260e52
GitOrigin-RevId: 542df9a21adbd55d7118b03203342d6086b3b191
Change-Id: I3b98ec3e6c9297737a92fee0568097cdeee0a194