[roll] Roll fuchsia [fidl] Re-organize compatibility test directory
The goal here is to minimize the "differences" between the various
implementation. Obviously, there will still be some differences in
naming, build rules, file structure, and the code itself, but the ideal
is to get to a situation where all of the "incidental" differences
between any two binding language implementations are removed.

Some examples of this:
- Each implementation is moved into its own sub directory, with its own
  BUIDL.gn file, its own /meta directory for its manifest, and so on.
- Every binding binary is named 'main.[EXT]`
- Every server has an identical name format
- Etc...

Note: due to the shuffling of some names, the CMX allowlist must be
updated as well. The modified files will be removed in a later CL in
this chain, one which this refactor makes easier to land.

Test: fx test fidl_compatibility_test
Original-Bug: 91343
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698994
Original-Revision: 807137e0e1302ec1cc02b5106c3279de30d90d7a
GitOrigin-RevId: 440eb9ff22bbf521d0fef45b084bf6f6c3ef453d
Change-Id: Ib5f42d93050cfb541309cbf005ab8db6396124fc