[roll] Roll fuchsia [fidl] s/server/impl/g for compatibility tests
Prior to this change, the various binding language implementations
exercised by the FIDL compatibility test suite were called "servers".
This is confusing, since the tests themselves involve spinning up two
instances of such a "server", one referred to as a "proxy" and the other
as (also) a "server". This repeated nomenclature is confusing.

Instead, the generic binary used to implement the compatibility test in
a particular language is now called an "impl". The two instances are
still referred to as the "proxy" and "server", respectively, though now
this latter term is much clearer and more unique.

As an aside: these tests may be much cleaner in practice if we stopped
"re-using" the same binary to implement both the proxy and the server,
which then relies on reading the incoming message's "server_url" field
to decide whether to forward or return. There is probably a better way
to do this, but that is quite a large change and out of scope on the
path to achieving the primary motivator for this CL: the CFv2 migration
of the FIDL compatibility tests. The bug https://fxbug.dev/104132 tracks
this future work.

Test: fx test fidl_compatibility_test
Original-Bug: 91343
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698995
Original-Revision: 7e9646483d67c84426aa8fa876b091afa7d1420b
GitOrigin-RevId: dcda708d3a58cedb911841c082955f74b016bb9d
Change-Id: Ie44aa399d96b5a79c465c2e31eb36132dc259a6b