[roll] Roll fuchsia [bt][hfp] Fix SLC Init three-way calling path
When the peer does not support three-way calling but does support HF
indicators, we were expecting an AT+CIND? instead of the correct
start to the HF indicators exchange.

Add tags to the logs for all parts of this component for ease of use.

Rename some of the SLC states to more accurately reflect the states we
are in, and add some comments from the spec for the triple-ending state
of SLC Initialization.

Original-Bug: 74091
Test: updated tests, manual test with peer

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/688452
Original-Revision: 5521e9f3be95aaabe95466edfd8165b32ca65925
GitOrigin-RevId: 51d96bf4bb6e3d7511bef4357f9c2cc644016dec
Change-Id: Ia2e7611363262e5addb3c90e5321d530e7222157