[roll] Roll fuchsia [ffx] Clean up stop message when nothing's running
The attached bug indicates that when we switched to the common
get_engine_by_name function for resolving emulator names, we missed a
case in "stop" when no name is provided. The error message from that
function was being discarded in favor of one about "deserialization".
While it's true the code couldn't deserialize an engine, this was not a
useful way to describe it to the user.

This CL adds a check for the case where no engine name was resolved, and
prints a separate message from the one when the engine is resolved but
fails to be deserialized.

Original-Bug: 103534
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/699854
Original-Revision: a8a46c0bdeb2f3a2813bddd9d8d170d86f566958
GitOrigin-RevId: a8fa9a4c48809729b3190773c20b4aeb74b353e5
Change-Id: Id235813ee798cc062dbf93ce10ad95f841178cf8