[roll] Roll fuchsia [ffx] Make log paths relative to PWD
Previously, if a relative path was passed to ffx emu for the log
parameter, it would be resolved by the emulator process, resulting in
the log file living in the emulator's working directory. This doesn't
fit with most developers expectations, as shown by the attached bug.

This CL explicitly sets the log file to be relative to the current
working directory of the FFX tool before launching the emulator, so it
will be in line with the user's expectations. Absolute paths are always
treated as absolute, so no changes are needed with the handling of those
paths.

Original-Bug: 103652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/699843
Original-Revision: 9684458d4fa0d99c00991c84fc5c5ca778040bc9
GitOrigin-RevId: e2e080fa75522994a7d6c11dad58f86a059e677e
Change-Id: I1709537070a3cebcfbf72377f8cba40fb3f4b0c2