[roll] Roll fuchsia [docs][performance] Change "Catapult" references to refer to "Chromeperf"
In Fuchsia we had started off referring to Chromeperf as "the Catapult
performance dashboard", or just "Catapult", but the name "Chromeperf"
is now preferred.

This only changes the references in the "user guide" page and the
files that link to that.  It doesn't change other references to
"Catapult" yet.

Add a redirect for the old page.

Add an explanation of the original use of the "Catapult" name.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698605
Original-Revision: 72e6e09831d7b365057d122a0d285648876e0630
GitOrigin-RevId: a7a8f53eb3505de8d40e207cca41a39f07474018
Change-Id: Ib0596c935251085aa05c1398b5eb1ab0054b25eb