[roll] Roll fuchsia [utest][debugger] Strip tag from register before reading at that addr
`test_memory_ops` will fail under hwasan because it attempts to read
from a tagged value via a call to zx_process_read_memory, which should
not accept tags under the address tagging ABI. The address we read
from is actually tagged from a segfault on a tagged pointer in a
separate function. In this particular case, the tag can be ignored
and the test should still be able to read from that pointer.

Original-Bug: 57112
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/699043
Original-Revision: b3de826c2ea2ee4c65349ef83200580703e3b332
GitOrigin-RevId: 5e19949935c9ad480e0bf3baa13aeee0da991dc8
Change-Id: I578c2ce2a4ebc57c59cfa4a17e1f101e73405616