[roll] Roll fuchsia [bt][avrcp] Report correct play status after time
When sending back from a GetPlayStatus command, the target would send
back the last computed play status, which often was just at the
beginning of the track, or the last time play / pause was sent.  Update
the current playback position based on the current time when
GetPlayStatus is called.

Update PlaybackRate to be a newtype for TimelineFunction and update the
related code.

Original-Bug: 102195
Test: updated tests, fx test bt-avrcp-target-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/699755
Original-Revision: 46ac21abe70b0bc0d86b708ffdb2ef03b2efe7ad
GitOrigin-RevId: d5c4e1f987b3a8df5eef6e923bd95a39cc85e4d1
Change-Id: I220ec5559c4e17874cc0926c5ebff3c559558295