[roll] Roll fuchsia [userboot] Svc stub passed to processes
An additional handle with name /svc and of type PA_NS_DIR
is sent to launched processes, this allows forwarding the other end
of the svc handle to a component at a later stage.

This would enable instrumentation of core-tests and component manager
(as part of a boot sequence).

Added test that verifies individual aspects of the protocol, including
the publish debug data integrity done through sanitizer_publish_data.

Due to C Bindings not being supported for fuchsia.io, we had to manually
decode the message, the bug with one shot decoding in LLCPP(not
involving an async dispatcher and a custom server implementation) has
been added to the manual decoding, such that it easier to cleanup
afterwards.

Original-Bug: 101831
Test: fx run-zbi-test early-boot-single-process-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/675469
Original-Revision: 8b5bd9c60ce428cec9424865dd24229cd9390b67
GitOrigin-RevId: 427d0a2deac3fda36b12ee3ead501f1ab553366e
Change-Id: I5cd34212abe040577596179f973df642f435d99e