[roll] Roll fuchsia [virtualization-net-tests] Adding fix for virtio net test flake
A previous change fixed a target issue, but left an issue on the host.
It is possible for the host to send a packet to the target before the
target's netdevice's receive buffers are ready. Add a loop to retry in
case this occurs.

Multiply: virtualization-net-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698667
Original-Revision: 0d013567d60b5fc9b1bf968be07d3d151acd54f0
GitOrigin-RevId: 8233525d32cb046f62b0ea8297f93bb23915517c
Change-Id: I570b5cb1f849013bdffee817b20be873468be935