[roll] Roll infra/recipes [gerrit_auto_submit] Pass variables around
Pass variables around instead of using members. Also, store state for a
gerrit host in the top-level of the state and under the name of the
host. If there's something under the name of the host prefer that state.

Original-Bug: b/238792726
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/699888
Original-Revision: 570386ec1bb77e35e3e4577fa715fd360ddf9caa
GitOrigin-RevId: e9e7968921aeb162c584fb9576574f77cc3a1a01
Change-Id: I8dcf935d28660fc2532687116af6ef04d92d863b