[roll] Roll fuchsia [ui] Add use_scene_manager_as_scene_owner flag
This flag enables developers to permute the UI realm between
scene_manager and root_presenter on products as we attempt to migrate
all products to using scene_manager.

When the migration is complete and we delete root_presenter, we will
also delete this flag.

Original-Fixed: 98781
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698977
Original-Revision: 7390996765db49996a600e4a8184435a1735c3f8
GitOrigin-RevId: c6b9e36f20b37cce8a901a38e194e0d4c91f41ff
Change-Id: I28909ecf13a3bdb02e34b926f7ea29e04d33ab19