[roll] Roll fuchsia [botanist] Stream blob writes from GCS to client
We previously copied blobs from GCS to disk prior to sending them to the
client. The client saw this as arbitrary delays prior to receiving a
response, and occasionally hit timeouts as a result. This CL streams the
blob from the remote bucket directly to the client, while also
simultaneously writing the blob to the cache for future requests.

Run-All-Tests: True
Original-Bug: b/238709034
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/700504
Original-Revision: ababf44818791f8eac9cbb44d41a89635f7e84a9
GitOrigin-RevId: 9441f5c2248db577c89657a0b3382948ac56a9f1
Change-Id: Ibb889c59375fc082e8df82053e9fa82790a15d08