[roll] Roll fuchsia [benchmarks] Run dart-fidl-benchmarks using Components v2 (second try)
This uses the Dart test runner component that was not available when
the earlier version of this was landed
(https://fuchsia-review.googlesource.com/c/fuchsia/+/672283).

The Dart test runner had a problem with granting access to the
custom_artifacts directory (see https://fxbug.dev/102330).  The
problem seems to have been fixed with a recent roll of the Dart test
runner, but we're not sure what fixed the problem.

Update the README.md by removing instructions for running the test
that are no longer applicable, because they pre-date the plumbing in
fidl_microbenchmarks_test.dart for reporting the results through the
performance testing infrastructure, and I don't think it makes sense
for each test to have instructions on how to run it.  These
instructions were originally in the Topaz repo under
tests/dart_fidl_benchmarks/README.md.

Original-Bug: 98667
Test: fx set terminal.x64 --release
  + fx test --e2e host_x64/fidl_microbenchmarks_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/687989
Original-Revision: fe93ef5fa2d0fd0eaadafc4e83398aa5cd9c16ae
GitOrigin-RevId: c3874131382d9a040dad692ecf8fcfc03f4627fb
Change-Id: Ibd18ddea7e74e1402360a540b845c2657e6dc7f8