[roll] Roll fuchsia [dhcp] Pass updated UpdatedAt to acquiredFunc
The previous refactor of DHCP client made it so that the updated
value of `Config.UpdatedAt` wasn't passed to `Client.acquiredFunc`
when it was called. This meant that the computation for the
valid-until address property for addresses assigned by the DHCP client
was wrong for all renewals (we were always adding the lease duration to
the time the lease was originally obtained, not the subsequent
renewal times).

Modified a unit test to act as a regression test for this.

Test: fx test --test-filter TestClientUpdateInfo netstack-dhcp-gotests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/700366
Original-Revision: 85842bd43f919755eed101fe1b3a1711fbe3a28c
GitOrigin-RevId: 091639549176ba392bf35c5d6be68a8d8895eb29
Change-Id: I408b88633042ce49b339f59023d979a11b2071ea