[roll] Roll fuchsia [ffx] Make config flag relative to CWD
Previously, if a relative path was passed to ffx emu for the config
template parameter, it would be resolved after the process had changed
the working directory, resulting in the inability to use files local to
the user's actual CWD. This doesn't make sense in most use cases, as
shown by the attached bug.

This CL explicitly sets the template files path to be relative to the
current working directory of the FFX tool before launching the emulator,
so it will always be in line with the user's expectations. Absolute
paths are always treated as absolute, so no changes are needed with the
handling of those paths.

Original-Bug: 95945
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/699890
Original-Revision: 95a0389dd69b5613b837c9bd000ece317176ce96
GitOrigin-RevId: 46c6f66a0dd470148970ec7f0c31aead3c7307b1
Change-Id: Ic115392fc15552823094efeb4a87f47eb9144af4