[roll] Roll fuchsia [archivist] Remove obsolete test
We already have coverage for this in the integration tests and it's
using the TestWithEnvironment fixture we want to remove.

Additionally the cpp-logs test is already using the structured backend
which is the default. So we don't need to have the same test twice.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/700786
Original-Revision: d37e3e056de2a33a34496880c88985576b7a8489
GitOrigin-RevId: ab274d184b85931314db90273e159a38fe194b83
Change-Id: I8d116fbe038e5383995ea99f0df386234aedcfbe