[roll] Roll fuchsia [governance] Move area descriptions into YAML.
This old structure was kinda confusing, with the descriptions living
in README.md and the metadata living in _areas.yaml. When the
"Graphics" area got semi-accidentally renamed, README.md didn't get
renamed, and so the description simply didn't show up on the API
Council page.

The downside is that the descriptions have to use HTML, rather than
markdown. However, this is already how the glossary works, so that's
not too unusual.

Also adds the API primary/secondary to the description for each area.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698330
Original-Revision: 8ed58ce39311a8561195f2562dc42df2c558685b
GitOrigin-RevId: b913991d06ba9523ebf1210b4c93ce709ca9fdcf
Change-Id: I0c692eea797414e63327bed9e90d714307e75a01