[roll] Roll fuchsia [dhcpv6] Abstract function for processing options
Add a process_options function which processes and validates options.
The function is passed the exchange type (valid combinations of
request and response message types), and therefore it is able to
perform validation specific to each message exchange, e.g. IA
options are not valid when receiving a Reply in response to an
Information-request message.

The Advertise message handler in ServerDiscovery state and the Reply
message handler in InformationRequesting state have been changed to
use the new function. The Reply message handler in Requesting state
was left untouched, as significant changes are being made in a
follow-up and usage of the added function is deferred until then.

Added tests to assert the new behavior.

Fixed a few tests around the stateless mode behavior where the Server ID
option used to be omitted, but are now required since we perform
validation.

Test: fx test dhcpv6-core-test dhcpv6-client-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/691943
Original-Revision: 0ca21f4a5d6effe8868f8c46a740ceebe6617644
GitOrigin-RevId: 0eab86569021e675ac6b2858486869caa0068ecb
Change-Id: I3e653a60c0c7a0f897521b80c3da3fee6c7aeec7