[roll] Roll fuchsia [dart][test] Immediately report when the real runner closes controller
This updates the dart wrapper runner to immedately close the client
half of the controller channel when running a non-test component and
the server channel is closed.
Channel closure in the ComponentController is semantically meaningful -
it indicates that the component has completed running.

In the original change, we didn't forward channel closure as we found
that fuchsia.test.Suite continues to run for a Dart test after the
channel is closed. Waiting for a Stop or Kill signal gave us a
signal which indicated when test_manager had finished draining
fuchsia.test.Suite. This had the side effect that calling Stop on a
non-test dart component would time out since we didn't forward channel
closure. To correct this case, we check that the component is a test by
looking for presence of an is_test=true argument in component start
info, and propogating the stop signal if the component is not a test.

Testing: fx test --count 1000 dart_realm_builder_unittests
fx test inspect-validator-test-dart against patchset 13 of
fxrev.dev/698139, which calls stop on a non-test Dart component.

Multiply: dart_realm_builder_unittests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/701143
Original-Revision: afef70aa07c2825ae57b48536e7cc5697df2ed35
GitOrigin-RevId: 14bff60989d76727b52ea0dd3c744546a7c234b4
Change-Id: I701e9e400114cf758ba17c80dbfdb16b34566786