[roll] Roll fuchsia [ffx] Change timeout message to something nicer
Since the emulator landed in March '22, the most common statement in
chats and bug reports has been akin to  "My emulator timed out". This
statement comes from the message that's printed whenever the emulator
fails to respond to a network request from the Daemon, which can be a
symptom of a problem, but can also mean the process is taking longer
than expected. It doesn't tell the user anything useful, and it doesn't
provide any information useful for debugging.

In this CL, we replace the "timed out" message with something more
useful for both the user and the devs helping them troubleshoot. The
wording around "timing out" is removed, and a more explicit message
about network requests is put in its place. Following that, we list
several of the pieces of information that we commonly ask users when
debugging these situations, including the emulator's networking mode,
acceleration mode, and pid, as well as common next steps the user can
execute on their own before reaching out for help.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/701230
Original-Revision: 8dc56a3b4c7b731d507ec32db607ff5eb5379cd8
GitOrigin-RevId: 7e15dce6cfee76d52ff6f03dcb8e2ec99df2fb56
Change-Id: I00633771c6a03d3d55e87554d26874655bdd39f3