[roll] Roll fuchsia [fidlgen_banjo][rust] Fix new compiler warning
Some old code was formatting a named variable but not reference the
name in the formatting string. This change adds the identifier to the
formatting string, but does not affect the behavior of the formatting. The new compiler warning was introduced in https://github.com/rust-lang/rust/issues/98466

Original-Bug: 104577

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/701186
Original-Revision: be4d5061154644cb8071be32b90ced4457e38f67
GitOrigin-RevId: 9a489709641b81e3069f8b6e71fac745327964fa
Change-Id: Ib6659729c9780b7a62af8d9a0314ef600f30767a