[roll] Roll fuchsia [fdf][compat] Improve devfs export log messages.
Previously we would log "Created <path>" prior to actually exporting the
device, and if exporting failed we would log an error that didn't
mention which device failed to be exported.

This CL changes it so that we log "Created <path>" *after* export
succeeds, and "Failed to export <path>" if export fails.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/700963
Original-Revision: 7ec4cbd17b18c06c6d167a2ecc20dd30ebd4fb21
GitOrigin-RevId: 87d23086d9931769af46114338a54b357176186b
Change-Id: I7c960c24934b1090a303307af9b457ce120f66e7