[roll] Roll fuchsia [security][sc] gn templates for structured config goldens
This change structures the current strategy for verifying
structured configuration against goldens by providing gn templates
for the "verify against golden based on build type" pattern:

- verify_file_by_build_type(): Verify an arbitrary file against
  goldens that may differ by non-eng/eng build type;
- verify_structured_config_by_build_type(): Verify a structured
  configuration against goldens that may differ by non-eng/eng
  build type.

The one use of this pattern is updated to use one of the templates.

In addition, the `--eng-golden` parameter for the underlying script
is made optional because verifying the eng build type is usually
not security-relevant. Finally, a README is introduced to describe
the function of files in `//src/security/policy/build/`, including
those used for the above-mentioned pattern.

Original-Bug: 104220
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/700078
Original-Revision: fd9a17f1621a81ed9ac4000139315efc7ac77954
GitOrigin-RevId: 53c1051f1ea6f950240743cfa6dd29db24704c48
Change-Id: I1dc9a25da9943ba3e1db37b9662e146f8c270faf