[roll] Roll llvm-project/libc [libc] Fix API for remove_{prefix, suffix}
The API in StringView.h for remove_prefix was incorrect and was returning a
new StringView rather than just altering the view.

As part of this, also removed some of the safety features.  The comment
correctly noted that the behaviour is undefined in some cases,
but the code and test cases checked for that.

One caller was relying on the old behaviour, so fixed it and added some
comments.

Tested:
check-libc
llvmlibc
libc-loader-tests

Reviewed By: gchatelet

Differential Revision: https://reviews.llvm.org/D129950

GitOrigin-RevId: adb2b7625e2f42760904ed43a37ecf83a2973286
Original-Revision: 85852da01d1505a64aec25e224ae5a230dc2183b
Change-Id: I5aeff53ae139aaa58d1fcbb01d4c31d0f420ae9d