[roll] Roll fuchsia [early-boot-instrumentation] Wire up SvcStash
The logic for doing so has already been added and unit test in a
previous patch. To keep things cleaner, this patch does the wiring
and introduces an integration test.

Provider component now also fills up a fake SvcStash and serves
SvcStashProvider which is offered to its parent.

The test component(parent) routes this to the child
early-boot-instrumentation component, which in turns offers the
profraw capability to the test.

The test itself just checks that the files are added as expected.
Provider will add the publish request to the SvcStash.
The first will have its token close(static) the second one will
still be open by the time the test runs, so its considered dynamic.

By adding the routing and logic into the main application we should be
getting instrumentation of the system component manager.

Original-Bug: 101831
Test: fx test -v -o early-boot-instrumentation-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/700506
Original-Revision: de7e2b91c052d7958bd0b2aaf9152fe52ee869d9
GitOrigin-RevId: 42c38c7ddf0182ef591cab42722213fb922c923b
Change-Id: Iad2747ebbe80ff790ed1ef6043ec1c2eda3a38f5