[roll] Roll fuchsia [ftl] Add diagnosis for running out of bad blocks.
Reports a failure symptom for when the number of marked bad blocks on an
ftl volume is at or exceeds (though it should never be able to exceed)
the number of spare blocks. Once in this state another bad block cannot
be marked bad and the ftl will fail out when attempting to do so. It is
the likely cause of failure when run on the device for an image that
otherwise appears intact.

Tested: On a known failing image.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/701557
Original-Revision: 160335006532b08ce0817a07561b229a98e9b40c
GitOrigin-RevId: 807ec71413b463387373709fe8199658c5413297
Change-Id: Ifde5369ce9057d5f7e164d43abe6ee6edf5bf73f