[roll] Roll fuchsia [perfmon] Propagate zx_status
We drop the zx_status_t we are given from the perfmon controller and
return it as a bool. This means that we don't have a way to determine
_why_ perfmon failed to start if it did except by reading the logs. This
changes the status results to now be a zx::status so that we keep and
error and are force to check/propagate it which should be a bit safer.

This also adds an integration test to ensure we can start and stop
perfmon if available.

Original-Bug: 101746
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698404
Original-Revision: 065242362c0d7e96aab00892e950383665691f3f
GitOrigin-RevId: 8f3221ed4d772380f7054117276f82cbf3b7ec54
Change-Id: I91805339cbfbbf7b9a42d101d332c8d46818dcb2