[roll] Roll llvm-project/libc [libc] fix strtofloatingpoint on rare edge case
Currently, there are two string parsers that can be used in a call to
strtofloatingpoint. There is the main parser used by Clinger's fast path
and Eisel-Lemire, and the backup parser used by Simple Decimal
Conversion. There was a bug in the backup parser where if the number had
more than 800 digits (the size of the SDC buffer) before the decimal
point, it would just ignore the digits after the 800th and not count
them into the exponent. This patch fixes that issue and adds regression
tests.

Reviewed By: lntue

Differential Revision: https://reviews.llvm.org/D130032

GitOrigin-RevId: 2dea111787923f67330397c3ae9d87d736fc8925
Original-Revision: 55b2f73b3d6a5ce963423287886351c7ebc674af
Change-Id: I3bce0726f253bd5bd1ecbc064586746f84e2593f