[roll] Roll fuchsia [acpi] Make device DAG reflect ACPI tables.
In I8fb14cae5e7bbbb05bb5ef58d8272336558c05fb we made all devices
children of the root ACPI device to work around the fact that we cannot
add children to a node that a driver will bind to.

It is possible though for multiple devices with the same name to exist
at different points in the ACPI namespace, and recently we've started
noticing this. This CL reverts to the state prior to
I8fb14cae5e7bbbb05bb5ef58d8272336558c05fb but continues to respect the
driver framework's constraints by adding a new `-passthrough` child
node to each ACPI device, which is what drivers bind to.

Test: boot QEMU, Atlas and check hardware works.
Test: check Atlas on DFv2.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/694727
Original-Revision: 689e98d8425afe2221466a1937c75d09629c3a1f
GitOrigin-RevId: 514b69490ea2d8f4bf6388e123b1fbf6a167cf1f
Change-Id: Ic3920bb471e2120ce2ecbc80b800a78bff0c771e