[roll] Roll fuchsia Reland "[cmx] Delete fuchsia_modular_testing"
This reverts commit c48fe4b5ba36fede147a94c5ceba7d450f334e05.

Reason for revert: Once again but with feeling.

The original change was reverted because an OOT roller expected the
package files to exist, and failed when they were removed from the SDK.

cl/458565033 removes the reference to fuchsia_modular_testing from
the roller, so now this package is safe to delete.

The previous attempt to remove the reference in cl/460354197
was rolled back because it also removed the fuchsia_modular package
that was referenced elsewhere. We are now removing one package
at a time:

1. Remove OOT (cl/458565033, submitted)
2. Remove from SDK (this change)
3. Remove the next package, fuchsia_modular:
   cl/460540399 (submitted) then fxrev.dev/699044

Original change's description:
> Revert "[cmx] Delete fuchsia_modular_testing"
>
> This reverts commit 0eb996ed5ea4bfc0b7103a6db88be7af6c1930ff.
>
> Reason for revert: b/238611056
>
> Original change's description:
> > [cmx] Delete fuchsia_modular_testing
> >
> > No longer used.
> >
> > Original-Fixed: 103687
> > Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/697127
>
> Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/699246

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/699573
Original-Revision: 2ab32853fd642ef8ac26674b4dbbb608edac7e44
GitOrigin-RevId: 6242944c145ac962fd6a86ce400e528016f5d808
Change-Id: Iba66acac0adefbec51823df296ba454605cc123c