[roll] Roll fuchsia Reland "[audio][drivers][hda] Changes to deprecate intel_hda::Status/StatusOr and use zx::status instead"
This is a reland of commit 4d50a61614359237dc021bac87b59f39ba272a65

The fix for fxrev.dev/696902 exposed (in the asan build) an address access issue caused by race condition between the Audio driver and the DestructWithQueuedThread test thread of intel-hda-controller-test.
The DestructWithQueuedThread and Audio driver use an object of type DspChannel. The test races to destructs this common object while the driver is still using it.
One way to handle this issue, is to call Shutdown() of the object to gracefully signal the driver, wait for the driver thread to join and finally call the reset() to destruct the object.
But this change in code will make this test redundant as  ShutdownWithQueuedSend tests the same sequence.
This new CL removes the DestructWithQueuedThread test as its no longer
needed.

Original-Fixed: 86895
Test: fx test intel-dsp-stream-test intel-hda-stream-test intel-hda-controller-fuzzers
      intel-hda-controller-test intel-hda-test intel-hda-utils-test-test
Test: signal_generator on Atlas
Test: intel-hda-controller-test under asan repeated 9999 times

Original change's description:
>
>Fixed: 86895
>Test: fx test intel-dsp-stream-test intel-hda-stream-test intel-hda-controller-fuzzers
>      intel-hda-controller-test intel-hda-test intel-hda-utils-test-test
>
>      Test with signal_generator on Atlas
>
>Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/696902

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/701687
Original-Revision: d2339fa8197bb4783a68b5593754a76fa10f9e75
GitOrigin-RevId: 245680fb0086a47c3e481d3c2eda98d15e2faf60
Change-Id: Ia6203fa503c700507a0fa33746161936fe475e6b