[roll] Roll fuchsia [ot-radio] Update pending tid log level
The ot-radio driver is tracking the pending TIDs while resetting, and
generate an error message if outstanding request haven't receive the
response when resetting the ot-radio driver.

This is helpful to triage issues related to reset in Thread 1.1
releases, since the force reset of a radio means the thread stack is
crashed and needs to brought up again.

However, when implementing the Thread 1.3 Backbone IF feature,
lowpan-ot-driver may be forced to restart when the Backbone IF is found,
which is a valid case and should not generate an error message. Hence I
am changing the log level from error to warning.

Original-Bug: 104100
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/701628
Original-Revision: 835b5ed5938b4bb852b5b6357c1b492e6beb1be8
GitOrigin-RevId: 252feb75b6e9a0a4f584517bf9a29cb1b7e5139f
Change-Id: If7ef93168321bcd342be84277e13744f68ca2303