[roll] Roll fuchsia [cmxtinction] Convert trace providers to CFv2
CFv1 is deprecated, time to move to CFv2!

This moves cpuperf_provider and ktrace_provider to CFv2. Instead of
having providers in a config file, we instead configure them as eager
children of the trace manager and route the the capabilities they need.

This also moves a few integration tests to use the gtest runner. It
makes some changes to when the tests bail out if there is not perfmon.
If we return EXIT_SUCCESS but haven't run any tests, we crash the test
runner. So instead, we start the test and immediately return.

Original-Fixed: 103220
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/693372
Original-Revision: 99a4477f063956fad954deb8e1c38d7cd781e3a7
GitOrigin-RevId: 02f63358a1048d453b065a9494b8de24cd828553
Change-Id: I62965ed226b258893fbd53e7c75ba175ddfdc3a4