[roll] Roll fuchsia [libc] Do not build lsan-test with the hwasan variant toolchain
Otherwise we will run into many undefined symbol link errors when
building this test because the hwasan runtime doesn't come shipped
with lsan. It could be possible in the future for this to happen,
but that shouldn't be a blocker now, so what we can do is just ensure
this isn't built with hwasan.

Original-Bug: 57112
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/700131
Original-Revision: e75705e8342a2541b6726434f85ef9e232c7a4ec
GitOrigin-RevId: a40d0ef435f96d2045bd6532b54330d2c0d9d4c1
Change-Id: I4ba84a207f5e5302ec74bd4ff49ef7edf882d32f