[roll] Roll fuchsia [ffx][pkg] Tests should specify abi revision
When the ABI revision changes, the package meta.far merkle (and
theoretically the meta.far size) can change. To avoid these tests breaking
when we update the ABI revision, this changes these tests to pin to a
specific ABI revision in most places. In addition, this adds a test to
check that we can still target the latest ABI, but we automatically compute
the meta.far merkle and size, rather than hardcoding the values.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/702142
Original-Revision: ef8ef485d75953b376ee301c0459e2ded8185054
GitOrigin-RevId: 501849cfb23ee932e75c26e846217cf8d2a371c7
Change-Id: Ic20d146355824708a835d28478783122d40222b0