[roll] Roll fuchsia [toolchain] Exclude hwasan variant from being mixed with linux toolchains
We hit a lot of build errors when attempting to build core.arm64-hwasan
because many linux binaries built for starnix would have a hwasan
variant of them built. This isn't surprising because we currently don't
ship the hwasan runtime for linux.

This fixes the issue by introducing a `fuchsia-only` tag that ensures a
variant will only be applied to toolchains for building fuchsia targets
and adding that tag to the hwasan variant.

Original-Bug: 57112
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/700075
Original-Revision: 50751984d604d17a9801d9733c87949c574aefea
GitOrigin-RevId: 98c60918d0f1b01b7791c6dccb9b016a3213ec6e
Change-Id: I988a1ec2dee436f84d4b296cc0a174c159f91504