[roll] Roll fuchsia [fidl][cpp] Add natural type support libraries in the SDK
Initially as part of the "experimental" category. The code generator is
already in the experimental SDK. Along with the natural type support
libraries, out-of-tree drivers will be able to build natural type
speaking FIDL bindings.

As the API stabilizes we'll upgrade it to "partner".

Original-Bug: 92356
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/702242
Original-Revision: 977a695c6bffd68123feb66a311a2c0263e59d5a
GitOrigin-RevId: be83b8c7ffed11ddee186d04427da746a5304794
Change-Id: I82c0da9191d76b20f2af83b68155b9e7751b563d