[roll] Roll fuchsia [f2fs] Add strict fsync for renamed file
This CL solved a consistency problem for renamed file.
The rename operation not only changes name of the file, but also
changes a dentry of parent directory. When SPO occurs after fsync, the
dentry change of the parent directory is lost.
To solve this problem, we manage ino of the parent directory as VnodeSet
and do checkpoint.

Test: fx test f2fs-unittest
Original-Bug: 102658
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/697882
Original-Revision: 724514ce076e1b68402e780cd816ba16793f46a3
GitOrigin-RevId: 89cfab9b8f2a28219735def31490b4c21180f239
Change-Id: I426d28ec0291c8081c60d3fa513121114970c749