[roll] Roll fuchsia [ffx] Generate ssh keys if none found
This checks the `ssh.pub` and `ssh.priv` configuration values, and if
no existing keys are found, a new ed25519 key is generated and saved at
the location of the first value configured. This is typically
$HOME/.ssh/fuchsia_ed25519 and fuchsia_authorized_keys.

Typical usage would be to check for the existence when initializing
a device via flashing/paving, or starting an emulator instance. This
is preferred since generating the private key when attempting to access
a device already initialized is pointless.

Also addressed issues from `fx clippy`.

Original-Fixed: 97603, 76032
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/699186
Original-Revision: 4ced5dd5fdfcd64f60e5a4404d9dc494e2dc9f10
GitOrigin-RevId: 31cf7f57dc936d82522bf8c2f69b72d5412e8f3c
Change-Id: I885bd0dfea54846e84314874a1ec6058bb34873a