[roll] Roll fuchsia [mixer_service] Add GainControlIds to PipelineStage
This functionality will be used to attach gain controls to a stage.
These will be used in MixerStage, where for each input source stage, the
gain controls will represent the gains to be applied, to be combined
with the gain controls that are attached to the MixerStage, to produce
the final mix output.

Original-Bug: 87651

Multiply: audio_services_mixer_mix_unittests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/701665
Original-Revision: c7cd9bb1aa8822119e78c6c823a48e2fe00c8d57
GitOrigin-RevId: 7c10ffa5a5604f73542f76f575c6cb5a0ef71b33
Change-Id: I86d9830d69148d387af08f059e45aed9625b686e