[roll] Roll fuchsia [mixer_service] Refactor GainControl
This change refactors the GainControl API to make it more compatible and
performant to use in the upcoming MixerStage implementation. On the
other hand, this does not impact the main functionality to interact with
and schedule gain changes.

A follow up change will be submitted to introduce GainControls, similar
to ClockSnapshots, to be added into MixJobContext.

Original-Bug: 87651

Multiply: audio-libprocessing-unittests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/701890
Original-Revision: 8e81a2ca1d2ebe56b5735e8347fc1fe2674eb5d9
GitOrigin-RevId: 2e581f646ecd436b3229be18e3421940960bd68b
Change-Id: Ife3549762f981ef8e1e404db3245ae3cb3005b71