[roll] Roll fuchsia [acpi] Add sleep state check to SetWakeDevice.
Ensure that a device is only set as a wake source if it is able to wake
the system from the requested sleep state.

Add tests for this and also add tests that power resources are turned on
when SetWakeDevice is called.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/701397
Original-Revision: 9362af9a604ec10fac39e957e94bca662325dce3
GitOrigin-RevId: 1ca61b40e59f831858671d5d090056ddf710a1b8
Change-Id: I5499e0884d004f0ec2a2e4a6219362dbc94d2863