[roll] Roll fuchsia [vfs] Add support for DESCRIBE for services
I added tests for this in the VFS library rather than a conformance test
because the conformance test does not currently support services and
it's not a change I want to take on for this bug fix. Services are also
only supported by a limited number of servers.

Original-Fixed: 104406
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/701395
Original-Revision: 3b7a7354af8c2e1104aae24946b8c145bb127eec
GitOrigin-RevId: 2f25c6e3a3e758e57830b7bd5b793b0f97ac52ac
Change-Id: I2c092ecf3421690e7c8d5c76ebada468dafd0e9f