[roll] Roll fuchsia [ffx][pkg] Add support for refreshing expired metadata
This adds the ability to refresh expired package repository metadata. It
does it by adding a notion of a "RepoStorage" trait, which allows
repositories that support it to commit changes to a repository.  It also
adds a "RepoKeys", which allows us to load private keys from pm-style
key files.

Original-Bug: 100773
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/701627
Original-Revision: 1953f917ca57635cddcc7c6a9a7868647e42da54
GitOrigin-RevId: 1582ea0a05bc8afc1b01b2d0af9b3d37a63f22eb
Change-Id: I3cd4fbc5a720ac1870456b16da1850ca1d60c7b5