[roll] Roll fuchsia [identity] Add AccountPreAuthState to AccountHandlerControl.
We want to avoid giving account handler direct access to the
device-encrypted disk (and therefore the ability to exfiltrate large
volumes of account data if it were to be compromised). Previously we
solved this problem by giving account handler access to stash but stash
is deprecated and therefore we need a new solution.

This CL adds a new opaque data structure to the `AccountHandlerControl`
protocol, letting the account manager store state on the
device-encrypted disk on behalf of account handler.

Further work will be required both to persist this data structure in
AccountManager (fxbug.dev/104514) and also to use this data structure in
place of stash in AccountHandler (fxbug.dev/104516).

Original-Bug: 104515
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/700919
Original-Revision: 1f5fca4a1b9ce8a7a8c665a3abe11c35e6a9dc83
GitOrigin-RevId: 90fcca5eac4b419e3181e27d0af5246c68517e19
Change-Id: Iafda5260b987d7c1901e21310a3cc2d06e1b3ebf