[roll] Roll fuchsia [starnix] Allow to create the init task without a filesystem
Most filesystem operation will require a Task to allow for permission
checking.

Before this CL, a filesystem was needed to create a task, which creates
a bootstrap problem.

This CL allows to create a task without a filesystem, and setting the
fielsystem later. This should only be used when creating the init task.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698382
Original-Revision: 07e8132686144fb7a47720fd3c6cf190a93f50a0
GitOrigin-RevId: 1c6b3baefef57c117acb5ff63c12adfdb311ec1e
Change-Id: I2f33b737aa60fa1a2b66b77cdf446009c8a1067b