[roll] Roll fuchsia [kernel][vm] Fix commit and resize race
If a commit is blocked on a page request and a resize invalidates the
range being waited on, the commit loop should not retry with a now
invalid range when unblocked. This used to work correctly until
fxr/693384 refactored the commit path and failed to bubble up the
termination condition from a lambda to the outer control loop.

The added test fails a range check debug assert without this fix.

Test: core-pager-writeback
Original-Bug: 105012
Original-Bug: 63989
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/703346
Original-Revision: dfaa5870ec54c66825902e7affb9a12ee7840949
GitOrigin-RevId: 792d3cc1e53910602c45c4b4aeb7672c51ab8345
Change-Id: I84ff3e1d2625664b9addcefe9b7c8d70055c273e