[roll] Roll fuchsia [fidl][cpp] Improve wire/natural conversion
Use the tricks from [1] to print an error when the user passes an
unknown type to the converter functions, as opposed to defaulting to a
no-op conversion.

Unfortunately that means we can't rely on the no-op template base case
and have to generate slightly more code to specialize the conversion
templates.

Add support for converting protocol endpoints and test them.

[1]: https://artificial-mind.net/blog/2020/10/03/always-false

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/702527
Original-Revision: 1e7fcfd3f44fe01dcb6003be033c33451a629f3e
GitOrigin-RevId: 006ffea660307904a650f6c7246d684bfe1208e9
Change-Id: I22d5f5d23911a5532eaa2f34206bbf3c9295d398