[roll] Roll fuchsia [kernel][vm] Handle gaps in DirtyPages
We can encounter gaps in DirtyPages if we raced with eviction and a
clean page was evicted before the DIRTY request could be resolved and
the page marked dirty. Fail the DirtyPages call in this case and let
waiters of the request retry looking up the page.

Test: core-pager-writeback
Original-Bug: 63989
Multiply: core-pager-writeback
Run-All-Tests: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/702322
Original-Revision: afa1a043164372192cfae505a42d2dff287b54af
GitOrigin-RevId: 2037638ae2ea0c8af5e63f7194784c3c053eb3b0
Change-Id: I7795e6fee7554a9f3742d1b34d8bacfb69f47c36