[roll] Roll fuchsia [kernel][coresight] Implicitly relativize access to component registers
Previously during a ROM table walk, we would access component registers
explicitly relative to the ROM table base. This prevented the (clean)
introduction of component class abstractions to facilitate that access,
as it would be hard to portably spell that offset in a way that would
makes sense when the component object is outside of the context of the
base ROM table; this also required a parallel `GetAt(uint32_t offset)`
API for all of the common component registers. In this change we reduce
that complexity by relativizing I/O to the current offset for each walk
node visited.

Test: k coresight walk on an Astro
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/437597
Original-Revision: 929c773cad85175b0061ff4dc4b31319b6feee83
GitOrigin-RevId: 8b2e7a88641ac082630a40880c76ed758859fe0f
Change-Id: I07b86f76804be55fded23bb4324cc60078066d3b