[roll] Roll fuchsia [netstack3] Move NUD to IP device layer
A followup will refactor ARP to use the NUD abstractions.

Note that this change removes a lot of unused code that was introduced
when NDP NUD was implemented. Later changes will implement NUD properly.
We take on the cost of throwing away the code so that NDP and ARP
based neighbor resolution for IPv6 and IPv4 can use a common
implementation sooner than later. This allows NUD to be implemented
for both IPv4 and IPv6 together - making sure future abstractions
are valid for both protocols.

Original-Bug: 99830
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/703153
Original-Revision: 9ff064b30848159a70fce9f5026d148338657afe
GitOrigin-RevId: 20dc1a187ff8b5f69694d15e231d1060bda6c400
Change-Id: I2620cac3febbff2c69f191c6e75e8ab80c94435e