[roll] Roll fuchsia [console-launcher] Retain dlog's virtcon handle
The current setup between dlog and console-launcher is inherently
error-swallowing. That's because dlog takes a handle to the virtcon
terminal as its stdout, and when that handle closes, virtcon gives you a
black screen. Retaining that handle allows us to continue to see dlog's
output after it dies.

Add logic to restart dlog if it dies, and do the same for the virtcon
shell, of which there is now only one. This same logic already exists
for the serial shell, and is reused here.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/702964
Original-Revision: bacfcb094098620c2c236f35f5534178f31e9dc7
GitOrigin-RevId: 6b312d9930dbed6cb1de64b66adb99943e164556
Change-Id: Ia72ce9129c252d9ff85b88a1419c116632f27fb0