[roll] Roll fuchsia [fidl][cpp] client side flexible method handling
Implement client side handling for flexible methods. This includes
sending the appropriate headers to indicate to the server that the
method is flexible and handling unpacking the result union used by
flexible methods.

This increases the complexity of the logic needed to correctly unpack
the result union. That logic was previously duplicated across
WireResult, WireUnownedResult (multiple times!), and
driver::WireUnownedResult. This was very error-prone as the
logic/templates needed to correctly unpack the result union (if any)
could be updated in one template without updating another template, and
there was a risk of having bugs in some code paths but not others, which
was very difficult to test for, especially since the architecture made
many of those paths very hard to exercise in unit test situations.

To deal with this, I created a BaseWireResult type which provides the
storage used for the wire result, the Unwrap and related accessors (if
applicable), and the logic for extracting the return
value/error/transport error from the result union. This ensures that
there is only one copy of that logic so it is the same for all
WireResult types.

Original-Bug: 88366

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/692530
Original-Revision: 096d81efe903b8b3a5d800dbcd690d1ee1669ed1
GitOrigin-RevId: 062c6aab6312bc0756b62de6cb07885b0a60c164
Change-Id: Ie6390cf0ccf4565af79200006b357ce1fc109fc2