[roll] Roll fuchsia [virtualization] Implement SET_SCANOUT, FLUSH
Implement a scanout on top of a flatland view that will be presented
using GraphicalPresenter. For this implementation we still use a
hard-coded size for the virtual display and we don't yet attempt to
resize this to match the size of the view.

We disable the GetDisplayInfo device test since this test is now coupled
to having a functional flatland service. This test will be re-enabled
when we can properly create a view within the test.

Original-Bug: 102870
Test: fx test virtio_gpu_unittests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698090
Original-Revision: 699f23b8ad133462e46f90039a79034e6d965550
GitOrigin-RevId: 28d8d4cf0626af59a4c993026146758dc1e2f1b8
Change-Id: I492399bab169006ca1c74fb27faa5f4c9979ffe6