[roll] Roll fuchsia [libc] Add trampoline for __hwasan_add_frame_record
Use trampolines similar to what's used for sancov functions for
circumventing PLT calls to __hwasan_add_frame_record. This is needed
in dynlink.c for instances where PLT calls cannot be made. The
strategy here is to define a local stub that __hwasan_add_frame_record
keeps a weak reference to so the linker elides the PLT entry and
resolves calls to the local stub.

This is also needed in allocate.c for a slightly different reason.
By the time we reach `map_block` for the first time via
`__init_main_thread`, `__hwasan_add_frame_record` will resolve to
the definition in the hwasan runtime, but since shadow call stack
is used in the runtime, we would page fault by accessing x18 which
is not setup yet intil after the call to `__init_main_thread`. In
allocate.c the stub is left empty to avoid a call to the runtime.

This CL includes some refactoring for reusing the sancov stub macro
for hwasan.

Original-Bug: 57112
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/702145
Original-Revision: c4bc7a08501cc35492ce87804afcae82a81f4e52
GitOrigin-RevId: 3eaf14a897763195003f47f24be9061888e7bd29
Change-Id: I0b56dd17df6cbca203bc1b8738691bba9f3e0668