[roll] Roll fuchsia [ui][input-pipeline] gestures: convert positions to millimeters
Before passing touchpad contact positions to the recognizers,
convert those positions from the units specified in the touchpad
device's descriptor, to millimeters.

Note that this implementation parses the device descriptor for
every touchpad event. If this proves to be a performance bottleneck,
we can add a cache later.

Along the way:
1. Plumb the touchpad contact units up from touch_binding to
   gesture_arena.
2. Update existing tests to provide units for the touchpad
   contact fields.
3. Add a method to read out the last event passed to a
   StubContender.
4. Add some logging to help dogfooders confirm that they've
   opted in correctly.

Original-Fixed: 105018
Test: fx test input_pipeline_lib_tests
Test: manually verified that both a single flick across the screen,
Test: and precisely pointing at a small target, were possible
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/703529
Original-Revision: 0992582510dab512de9a5b0978aa952b1aedd246
GitOrigin-RevId: a55a60bcd9693fc2555198b9b605dc333a751e47
Change-Id: I15bee507dcea4d88001cc925edf35b46e8698b4c