[roll] Roll fuchsia [ffx][package] Fix ffx package build --depfile
According to the [gn reference], all files that are read by the action
must be included in the depfile. This changes `ffx package build
--depfile` to make sure we include the files we'll be putting into the
far package.

In addition, our gn will expresses that `ffx package build` produces
`package_manifest.json`, so we don't need to include that in the depfile.

In addition, this makes sure that the entries in the depfile are sorted
and deduplicated to make sure we generate consistent files.

[gn reference]: https://gn.googlesource.com/gn/+/master/docs/reference.md#var_depfile

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/703722
Original-Revision: 13d41b0444fe65bb2c6e93eb797e7b5d52b21959
GitOrigin-RevId: f8550c487a630afc908045aea5e431e7d7a53535
Change-Id: I793d0260aff15a463cf746b2c0c01a3d38c23d5b