[roll] Roll fuchsia [ffx] Launch an editor from ffx emu
One of the biggest issues with the initial version of `ffx emu` is the
difficulty users found with modifying configuration. We put forward
go/ffx-emu-config-design as an improvement for this situation.

This CL implements the most crucial aspect of this design, which is
actually launching the editor on a launch configuration object. We do
this via an "--edit" flag, which starts up the user's default editor
(via the $EDITOR environment flag) or '/usr/bin/nano' by default. The
resulting value overwrites the existing FlagData object for the emulator
engine, as long as the editor process returns a successful error code.

Original-Bug: 98268

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/701233
Original-Revision: d213884466e5fca248d11264e32c0ea357b3ac9d
GitOrigin-RevId: 7a4ac39d669ee3669b7050e5ce2677a67bdc4580
Change-Id: I38efa40b8ba53452a911cd52d5d8592bb462ecf9