[roll] Roll fuchsia [kernel] Add restricted aspace to process dispatcher
This is the address space that will be used by threads that are running
in restricted mode.

Also moves the normal aspace to the shareable process state, since this
will be shareable in the future (the restricted aspace will not be).

This change does not introduce any way for clients to create processes
with restricted address spaces, or use it in any meaningful way.

A follow up change will add the ability to create processes that both
share their "shareable process state" with other processes, and contain
a restricted address space.

Original-Bug: 91399, 98922

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/673428
Original-Revision: 2afc00b85255cb1b6dfdf33ece0e08beff6d42df
GitOrigin-RevId: 89f00a1ac78274769b14eff76ccf2fbfdc435ab6
Change-Id: Id8ecac087378816f27fe0abd88f4c447ef527cc3