[roll] Roll fuchsia [component][fuzzing] Consolidate engine main functions
This CL extracts the common parts of the realmfuzzer and test engines
into a function templated on how to create the runner. Notably, it does
not update the libfuzzer engine: the code that integrates its engine,
runner and controller provider is missing! That omission will be
addressed in a subsequent CL.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/695590
Original-Revision: 06779b8a5e64a2ec61335fb8cc2dfac97dd96007
GitOrigin-RevId: 670a53721be1611cc72ec44f0c8763bb15b03959
Change-Id: I4f0f83269bf24847342707774c5945acc54b2959