[roll] Roll fuchsia [a11y] Announce selected and toggled states for buttons
Note that there's a fair amount of cleanup required in the
ScreenReaderMessageGenerator class. I chose to leave this cleanup as a
TODO, because it's likely we'll need to cherry-pick this change so I
wanted to minimize risk.

Original-Bug: 105052
Test: fx test a11y-lib-tests, manually verified on device
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/704144
Original-Revision: 8d72afba381bc9b88f1497b9bad3dfe9a1d9c821
GitOrigin-RevId: fadf9983e20eb26f0c70b0898bffd44cdd3aa030
Change-Id: I6bcf27b9585bb263a3f3e5bc6cb40fa55af0b115