[roll] Roll fuchsia [docs] Document the big-picture for the configuration of components.
This is the first publicly available version of a family of documents
that we've developed and discussed internally for a while.

The structure of how this integrates with docs is based on a very
useful discussion with Nick - thanks! Adding the reviewers he suggested
along with the people who reviewed in its most recent doc form.

Original-Fixed: 95374
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/701246
Original-Revision: d001cd0df4115aaad2abcde78be7fac3e35a968e
GitOrigin-RevId: 118f2c9a3e9d779d0e7e07e2a66847c298c75660
Change-Id: I744c10f9495286eafa7bce1a9853591a8d4ed958