[roll] Roll fuchsia [fidl][server_suite] Top-level structs, tables, unions
This CL adds tests for top-level structs, tables and unions in requests
and responses.

For tables, where there is optionality, bindings may or may not check
for the presence of the value. It is not necessarily important to
require checking for presence / absence in the binding because the
test runner will ultimately check the correct presence bytes are
received, so no mandate will be issued for untested behavior like
this.

Original-Bug: 102339
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/704014
Original-Revision: c7bc0a23474844add6d7a2c4ed1edd139f66ba67
GitOrigin-RevId: e923cc489844bf1261e941813dd49b9859db9308
Change-Id: I80db1af556b72b36012b85a0f96174ac557cc7f8