[roll] Roll fuchsia [kms] Remove the CFv1 component from the build
KMS is not load-bearing in its current form, and is currently blocking
the CFv1 -> CFv2 migration of `tee_manager`.  Removing the component
definition causes no functional or coverage regression, but does help
unblock the v2 migration for other components.

We keep the code (and unit test) in-tree to be referenced and
potentially reused for TPM-related work and can make a decision whether
to bring a `kms` component back or remove it later.

Original-Bug: 103105
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/704407
Original-Revision: 98a116e49bef96bef045054e6a7af599d41b3050
GitOrigin-RevId: d582ce745d134fdb6f56470573b6cb2871c62459
Change-Id: I250811b9b694d0a6b289c189efa7e901f7b6e2a0