[roll] Roll fuchsia [dfv2] Delete OwnedOffer
Now that we do not re-generate offers during the lifetime
of a node, we can remove this struct that stores an owned
offer.

If we do need to make a deep copy of a fidl table, convert
it from wire format to natural format and then back.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/704048
Original-Revision: c6af865509a3fb92adfc5e216795a137df726416
GitOrigin-RevId: 47b443ba65078701f98367421b3d4c0298e8edd2
Change-Id: I9e910e04fffe8dc96391a0fa9b091053e7f650b5