[roll] Roll fuchsia [fidlgen] Namespace HandleSubtype values
The current values to the HandleSubtype 'enum' are not namespaced and
are exported with ambiguous names like `fidlgen.Channel`. It is clearer
to prefix these with the enum name ("HandleSubtype"). This change also
eases the upcoming introduction of a type to mirror the IR for resource
declarations (a more natural fit to `fidlgen.Resource`).

We also s/HandleSubtypeHandle/HandleSubtypeNone, as the latter is
clearer and more precise.

Original-Bug: 7660
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/704242
Original-Revision: d290797438dbd34a0ac0b406d14f036dd2b21bac
GitOrigin-RevId: 868bcaa3d861003ee162b7525c7f058a33f6f4bb
Change-Id: I27a4d4f9d644421e537cd859fa5a045a8318f4ab