[roll] Roll fuchsia [pwrbtn-monitor] Fix test flake
Fix issue where test would flake if the first async routine to
connect to the power statecontrol service was the OOM routine.
When this happened the disconnection of the routine caused the
oneshot channel to be dropped, making the test complete too early
and leading to errors.

MULTIPLY: pwrbtn-monitor-integration-test: 500

Original-Fixed: 103485
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/704537
Original-Revision: 4dbc559330147f3ff41436a5c8cb45125b9b9720
GitOrigin-RevId: 0fcf80f68e466fdbde6dc0db9fb3c79e566241a9
Change-Id: I5df0302e6905f660a4658ff9c9eab00abd60a6ee