[roll] Roll fuchsia [fidl][cpp] Client support for fpromise promises
Today fpromise is the C++ promise library used in production. This CL
adds a helper library to transform callback-style async FIDL calls to
promises.

Because there is a new C++ fasync futures library on the horizon, this
CL opts to keep fpromise support in a separate library instead of built
into the C++ bindings. We'd like to reserve the built-in support for
fasync.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698173
Original-Revision: ccc710127e829ce50615fe1ea263d2582d5e86fd
GitOrigin-RevId: 3f45ea2bd34414cdf85f0656fb14687e868a0b12
Change-Id: Ie3a57ef27572e7170bf7e0cf1922de154daec543