[roll] Roll fuchsia [trace_manager] Fix Microbenchmarks regression
https://fuchsia-review.googlesource.com/c/fuchsia/+/693372 caused the
fuchsia_microbenchmarks syscall tests to regress fairly significantly.
This is due to an errant logging message being left in.

fuchsia_microbenchmarks has the ability to register as a trace provider
if you pass --enable-tracing, which we do. This means, every time we do
a fuchsia microbenchmarks run, it starts up and registers as a trace
provider. This causes trace_manager to emit a log, which then then spins
up archivist, which then spins up overnet, which all lasts long enough
to run at the same time as the the actual benchmark, which, even though
it's on different cores, likely interfers with the syscall timing.

Fixes: 104799
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/704942
Original-Revision: bf32682084421bcf69212cfabc78c7f3c686fe54
GitOrigin-RevId: eadf6f7338dad5f3766ef2b55c7dd6c2491cd0a1
Change-Id: If99d39c90249754292506324dc5ab349aa127afd