[roll] Roll fuchsia [netstack3] Add connect_udp_listener function
Add a function to replace an existing UDP listener with a connected
socket. This makes it possible for Bindings to implement POSIX
connect() as an atomic operation in Core. Making this an atomic
operation makes it simple for Core to propagate socket options when
connect() succeeds.

Original-Bug: 96573
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/662314
Original-Revision: 7d907395b7d1c9f1f4fa3d627f420b40421fb585
GitOrigin-RevId: ce2c07a365d5c566bbad012984bf0b3ec31e3b9a
Change-Id: I5dd16e888f6fd62fa51a49a36d12809179a45ff8