[roll] Roll fuchsia [test-ui-stack] Implement simple touch input helper
This change implements a very simple touch input helper component, which
will serve as the basis for the v2 input synthesis component.

NOTE: Some of the code here is copied from the v1 input-synthesis library.
We are forking that code consciously, as it will be the basis for OOT
input testing and SL4F.

Test: fx test input-helper-tests, fx test flutter-embedder-test-ip with
fxr/699870
Original-Bug: 104803

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/702985
Original-Revision: ae416694509e4495fb436f519966f5bbafb7a260
GitOrigin-RevId: ae3d7438fe17ddbe01fcbbb23a3b9f787205858d
Change-Id: I8e91aa2f24aef97733bf8c6ba7a8ff3aa2b212f2