[roll] Roll fuchsia [aml-cpu] Remove GetPerformanceStateInfo error msg
Removes the error message associated with an out-of-bounds state
parameter for the GetPerformanceStateInfo request. We expect to hit this
code path once during boot, so logging an error is not necessary.

During boot the power_manager queries the CPU driver to find out the
frequency/voltage information for each of the driver's supported power
states. It does this by calling GetPerformanceStateInfo for each power
state. Since the power_manager doesn't know how many power states are
supported, it queries the states 0-20 (MAX_DEVICE_PERFORMANCE_STATES)
until an error is returned for an invalid state (10 in this case).

Original-Fixed: 105250
Test: logging change only
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705205
Original-Revision: 28e750adcad104b7330cd042eafd3eefaf8a1447
GitOrigin-RevId: b8f874bb8ffdce7d59d0dce99f48bc420ccfc0d5
Change-Id: I4eb2237516b1e0b13b1421feb29a8c6e53fb1f10