[roll] Roll fuchsia [lowpan][openthread_rust]: Fix for UDP `ot::Message` leak.
It would appear that a previous assumption---that the UDP handler
method would take ownership of the underlying otMessage instance---was
incorrect. This is leading to `otMessage` instances leaking.

This change makes sure that we don't attempt to pass ownership, which
ensures that the `ot::Message` instance is properly cleaned up.

Original-Bug: 105401
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705268
Original-Revision: 24d257d7e061c5adcca8a41d3cc8fa58c6abf18f
GitOrigin-RevId: 36d973491ae93a81c9987e0ef3c2540bde20750d
Change-Id: I1ede86c2a3476d745b4aad153ac3b316bf7d0c0b