[roll] Roll fuchsia [gfx] Add Color Conversion API
Hookup the new color conversion API functions to GFX via the Engine. This allows clients of the ColorConversion FIDL API to call
SetColorConversionValues and SetMinimumRGB on the Scenic GFX backend.

The old method of using the GfxCommandApplier still exists in the code, and will be removed once this API is hooked up to SceneManager.

Original-Fixed: 96971

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/704028
Original-Revision: 0952d931321ff26022baca82ce547bf1ff5715e7
GitOrigin-RevId: a100ad1ff61077b364882712ada8bca74e492697
Change-Id: I069cfeddd3303beaa407bd1facec1419c3a0c85a