[roll] Roll fuchsia [component][fuzzing] Make generic fuzzing integration tests
This CL moves most of the realmfuzzer integration tests under
//src/sys/fuzzing/common, in much the same manner that most of the
Runner unit tests are under common/ and use macros to specialize for the
specific runners. This is done for the same purpose as well: to
facilitate creating integration tests that are the same for both
libfuzzers and realmfuzzers.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/695591
Original-Revision: ab061b26fcbc6f407ce6e3326ab087c9629d7218
GitOrigin-RevId: 19603749da67c45f40e8e84616867fe679ec041a
Change-Id: I2d0ae5a1b171c88821f9ead1d4780243afb1c062