[roll] Roll fuchsia [fidl] Client-side dynamic test suite
This CL introduces a client side dynamic test suite to FIDL, which is
meant to test edge cases of FIDL client bindings. It follows similar
patterns as the server side test suite, but tests are driven from
clients in bindings under test, rather than the harness.

A VERIFY_* mechanism is introduced that parallels ASSERT_*, but reports
errors to the test runner.

The Channel and Bytes classes are factored out of the server test suite
to a shared location, which meant not using test assertions directly in
that logic.

Original-Bug: 102339
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/700344
Original-Revision: cea0022b8a7b2ffecf2b178bc56c986b8f602c4f
GitOrigin-RevId: 99fa07b1787c07a95a0cd80ad3a843ab9a73bd74
Change-Id: Ib98070f32240916a469d5fb6fc7574886548c45b