[roll] Roll fuchsia [starnix] Correctly preserve execute permissions in fork
The bug: If the same vmo is mapped twice, once using a handle without the execute right and once with, and the mapping traversal found the
non-executable mapping first, the hashmap cache would get a clone of the non-executable handle and this would be reused for the mapping that was supposed to be executable.

The fix: delete the hashmap. It was intended as an optimization to avoid making extra clones, but it could never have done that in the first place, because you can't map a private VMO twice. The only entries retrieved from the hashtable would be shared VMOs, and you could have just cloned the handle instead of looking it up.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/701604
Original-Revision: 43a4aa0075e81a03d1ed8d590dd8daaff897098b
GitOrigin-RevId: a8c6a4cd981265e1cf744cf81ce44cde8ab8ee03
Change-Id: Idfb7af6b38be39f6d8092932e0ee81bc1abed757