[roll] Roll fuchsia [f2fs] Fix a corruption bug on a Page bitmap
The CL fixes a bug in VmoNode::CreateAndLockVmo() where an wrong
offset was used to calculate the range of Pages that kernel had
reclaimed in a VmoNode. It would result in a corrupted Page or
a page fault due to corrupting or overflowing on VmoNode::page_bitmap_
keeping the valid Page information for each VmoNode.

Multiply: f2fs-unittest: 200
Original-Bug: 105351

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/706222
Original-Revision: efc2b68842838185f9c79e46361eadfc6183979c
GitOrigin-RevId: cab3014fbed73e9af4c0a1edb8997ca789ccd65e
Change-Id: I0e48db212ed617fca49dc79fcc364dc473150c6c