[roll] Roll fuchsia [fdf] Add Device Group Manager to driver manager
Add classes for DeviceGroupManager, DeviceGroup, and
CompositeManagerBridge.

The DeviceGroupManager is responsible for managing
device groups.

The DeviceGroup is partially abstract class for
managing the state of a device group.

The CompositeManagerBridge is a bridge interface that
allows DeviceGroupManager to agnostically interact
with the DFv1 and DFv2 implementations for composite
handling.

Original-Bug: 103182
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/703684
Original-Revision: a3ddcbe3a30ef327d304d42ac5e9b30e68a75c14
GitOrigin-RevId: dfec23cf2bf7889c7b8085ee399a49dc4d6e6f17
Change-Id: I9a856efd168b8d81bd31f434c8a21d322a4e698d