[roll] Roll fuchsia [hid-parser] Limit the size of device descriptor
There is such a thing as allocating too many bytes for device
descriptors. The fact that we can allocate 2GB of memory for this
purpose doesn't mean that we should.

This change puts an upper limit to the allocation size to 1GB, I hope
that this would be enough to handle realistic usage.

Original-Bug: 64791
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705484
Original-Revision: 305bc499d56fb04499f8b85597ef580f91ef9315
GitOrigin-RevId: 07f3cdf28503f566016e4c768f2ed6fae22cae88
Change-Id: I374b2ece9c706932fa599792a90bb23ff388414c