[roll] Roll fuchsia [ffx][pkg] Add transparent gzip support to GcsRepository
When a file is gzipped compressed on GCS, GCS is unable to report the total
content length. To fix this, this changes GcsRepository to cache the file
into a temporary location so we can compute the total length, then return
the stream to the caller.

Original-Fixed: 105309
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/706103
Original-Revision: 16172b3c4a02aeeaeb9f09eb3072254289034553
GitOrigin-RevId: d8f556589462e383f395f6dccfd11a3abd42ea28
Change-Id: Ib0a371bfe04ca41f52129e341266ccd1c96c0699