[roll] Roll fuchsia [kernel][vm] Free failed pages in PhysicalPageProvider
If SupplyPages() fails, the pages left over in the source splice list
need to be freed. Fix the freeing logic which was previously freeing
from the page list that the splice list was created from, and would
therefore be empty after the move.

Run-All-Tests: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705625
Original-Revision: 53e997bedc37a12263b45fad992b86271518b704
GitOrigin-RevId: 1580ffdf7fb65a9298fee25dab559256328af9c0
Change-Id: I477415d710fe94dc0be6de4591edd9f6ef7e59c1