[roll] Roll fuchsia [intel-hda] Fix device_get_fragment_protocol
The last argument of this function is a void* that must
be the protocol struct. Passing a ddk:Pci here is not
allowed, even though it works because it wraps the
protocol.

Fix this by using the ddk::Pci connect function.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/706309
Original-Revision: 1ff3308b44437108dcc993d97bcfe3d2be0289b5
GitOrigin-RevId: c42a2815ebcf37d0859f6d35ec762ca35a382b49
Change-Id: I2c9c2f18da44a4da0546f638ffc8c362508713f1