[roll] Roll infra/recipes [checkout] Add option to allow skipping patching
We generally want a tryjob to fail if it's triggered against a CL in a
repo that isn't actually checked out by the builder. However, this is
not the case for SDK image subbuilds that build public images but are
triggered by internal SDK builders. In this case it's fine for the
subbuilds to ignore patch failures if the affected project isn't
included in the checkout, as long as the parent SDK builder is able to
apply the patch.

Original-Bug: 105507
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/705604
Original-Revision: 091e5e4f600f40ce6c44b566620c99a5555c94c4
GitOrigin-RevId: c33b10dce4404933d0f3da1d7e61d2e49200f4d9
Change-Id: I815a1e09bbea606eebefb68b2daa2a9f9385d0d1