[roll] Roll fuchsia [ntr] Add option to wait for any IP address
Some tests want to be able to wait until the interfaces they add to the
hermetic network realm have their autoconfigured IP addresses and have
completed duplicate address detection, so we add an option to
add-interface to wait for any IP address to be assigned to the interface
before returning. This is not strictly guaranteed to be the
autoconfigured IP address, but it should be unless the client is
intentionally assigning an IP address statically by separate means.

Original-Bug: 105295, 105506
Multiply: network-test-realm-integration-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705542
Original-Revision: ea575794ae2abadd723f1cb613b483dcb01fe4c8
GitOrigin-RevId: dcff4f5a5312e644184d79045dce5f1797a95df2
Change-Id: I29aad57cdbbcffe33336f9fd0adca21b578d3f04