[roll] Roll fuchsia [test_manager] Refactor to introduce new scheduler
This change is a refactor of test_manager that allows for the
introduction of new test suite scheduling strategies. This change
introduces a new scheduler module that contains the Scheduler trait.
This trait can be implemented by any object and dictates how a
collection of suites is scheduled and executed. This change also adds a
new enum to the facets module that indicates whether a facet is resolved
or unresolved. This will be useful when we add the ParallelScheduler
because the ParallelScheudler will need to resolve test suite facets
before invoking run_single_suite. We want to avoid resolving suites
twice so checking the status of the facet resolution in run_single_suite
will be necessary. Please refer to the following design doc for more
information - go/fuchsia-test-manager-refactor

Test: To test my changes, I made sure running "fx test
      test_manager_test" still passed all tests.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705599
Original-Revision: 6ecb890b2f0187e6e803c1add5c38b4689426901
GitOrigin-RevId: 4ac2aa17d3f1fd7c4f7889cd00ef0f28b89b203a
Change-Id: Ib5eb109cbca63db8e76e76d079973d419077b42d