[roll] Roll fuchsia [brcmfmac] Fix compiler errors when kALL filter is on
Turning on the kALL filter for BRCMF_DBG no longer causes the build to
fail to compile. The real fix for this problem is to add a build
target that requires the kALL filter to be on. This way we cannot
check-in code that fails to compile when the filter is on.

This change renames len to length_ in the macro expansion so there is
less chance of a conflict with variable used in the scopes where the
macros expand. (This is actually a generic problem with macros in C.)

Test: None. Debug code only.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/704528
Original-Revision: 43b10dbbb147f588c5585234c1568e67d0a2adf4
GitOrigin-RevId: 72830e9afac8029e29210c3155f4d24bb4f3a00a
Change-Id: I1250e891206923ad4becb9b04ada37aad0687882