[roll] Roll fuchsia [netstack3] Implement disconnect for UDP
Add support in Core and Bindings for the Disconnect FIDL call that
corresponds to POSIX connect() with AF_UNSPEC. This allows applications
to remove the remote address for a connected socket, turning it into a
listening socket.

No tests are added in this patch, but it should cause some out-of-tree
tests to start passing.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705186
Original-Revision: 9bcb159e4f7d6d57b6bb06775f64eef05da5ad69
GitOrigin-RevId: 2b9e980fe851d5e801cd8c54eac92d99b9ed047c
Change-Id: I3fc26891eb045519dddc3531f99ead2cdb6d3e53