[roll] Roll fuchsia [mediasession][examples] Convert example_player to v2
Convert the media session example player into a V2 component that can be
monitiored with the `mediasession_cli_tool`.

An `example-players` collection is added to the
`mediasession.core_shard.cml`: this shard provides the needed
capabilities from "#mediasession" to the example player V2 component and
the fx shell binary `mediasession_cli_tool`, which is used to
control/monitor the state of one or more example players. The
mediasession core shard is used instead of a dedicated core shard for
the example player because it already contains the needed capabilities
for both the example player and cli tool.

A README is added to describe how to configure Fuchsia and run the
example player component with the `mediasession_cli_tool`.

Original-Bug: 103877
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705035
Original-Revision: 89aa013b512e7c0cb3f6857891e31b2c96d00418
GitOrigin-RevId: 3c18840ad37abf83ca7529f5d50a66a234b68bc2
Change-Id: I1760e6352ef0f3c24d5499e6db517dd45f78f087