[roll] Roll fuchsia [benchmarks] Add initial allowlist for performance metric names
This adds an initial implementation of an allowlist and applies it to
a single test.  More tests will be covered in later changes.

The process for adding and updating an expectations file is not yet
very smooth.  I will improve that in a later change too.

sl4f.api file: In practice there is no API change because these parts
of performance.dart are not used via the Fuchsia SDK.
MetricsAllowlist is only made public so that it can be tested in
performance_test.dart.

Original-Bug: 105202
Test: fx set terminal.x64 --release
  + fx test --e2e host_x64/fidlc_microbenchmarks_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705628
Original-Revision: e76808b8b711758e9d7776a0625bdd39cbfeac4a
GitOrigin-RevId: f86d4e8530b0ebce5d4d8721033a3b254bf14461
Change-Id: I1c2cffdd208aa908c57603fd1a4649c27f6a0d1c