[roll] Roll fuchsia [build] Remove write_file() in sdk_atom().
This replaces a write_file() call with a generated_file() target
in the sdk_atom() template definition. The result are equivalent
except that the files can be generated in parallel.

Adding a new dependency to src/tests/end_to_end/fidlcat/BUILD.gn
is required to avoid GN complaining that a target uses an input
file without depending on the target that generated it.

In practice, this only touches 4 files in  a core.x64 build config:

- gen/src/zircon/lib/zircon/zircon_sdk.sdk
- host_arm64/gen/src/developer/ffx/plugins/emulator/aemu_internal.sdk
- host_arm64/gen/src/developer/ffx/plugins/emulator/qemu_internal.sdk
- x64-shared/gen/src/zircon/lib/zircon/zircon_sdk.sdk

Original-Bug: None
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/706022
Original-Revision: 3b89a61d7799f0cbe8e0b57fb5465f078be6a943
GitOrigin-RevId: 36819b1a3cea0c10a7a43ec98b61faf12c3821e1
Change-Id: I30a54c3fb804cf1c7d2d3f65cdd66a38777c0fc9