[roll] Roll fuchsia [netstack3] Add reconnect_udp function in Core
Add the ability to atomically update the destination of connected
sockets to Core. This is required to support operation of POSIX
connect(). While it could be implemented as remove-then-connect, pushing
the functionality into Core makes it easier to keep socket options from
one to the next.

Original-Bug: 96573
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/665883
Original-Revision: 4bbf3203b87f2cc778388dfe47023a484ffa6b10
GitOrigin-RevId: 375dd35eb966ca8a0200f0841249c211efd43012
Change-Id: Ie325acada67f53820d563e631bf61dcc4fbe32f1