[roll] Roll fuchsia [fidl][cpp] In-args: use `const T&` for value and `T` for resource
The natural messaging APIs have signatures like so today:

  Foo(fidl::Request<Foo> request);

this causes extra copying. This CL changes those to look like:

  Foo(const fidl::Request<Foo>& request);

for when the domain object is a value, and

  Foo(fidl::Request<Foo> request);

for when the domain object is a resource.

The completer APIs used looks like so today:

  void Reply(fidl::Response<Foo>&& response);

The `T&&` is very subtle:
- it forces the user to always std::move their value in
- the signatures suggests that the function consumes the response, but
  is actually not the case: nothing will happen to `response` unless
  `Reply` explicitly moves it elsewhere.

So that's changed to

  void Reply(const fidl::Response<Foo>& response);
  void Reply(fidl::Response<Foo> response);

as well in this CL.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/706303
Original-Revision: ff162a0a81c43a8a2e828cab741c662ac103c66f
GitOrigin-RevId: bf9d5d5d35928093766589fedf8bacfaeb873942
Change-Id: Ice90c5cf18206085aa76f660279d8aa8b6769d4e