[roll] Roll fuchsia [fxtest] Populate FUCHSIA_NODENAME for local host tests
Uses get-device-name to populate the FUCHSIA_NODENAME environment
variable for host tests. FUCHSIA_NODENAME is provided by infra scripts
when these tests are run in CI/CQ, but the variable is not provided
when tests are run locally.

Testing: manually verify that the nodename is available to a host test
that uses it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705736
Original-Revision: 195abd7042bba906a450a580a112265ba6b1439d
GitOrigin-RevId: 4c8ed7396e3e8ac6103987c4a77a8e4e5f7881ab
Change-Id: I87d72bfe26d61087afc9a396aef863dc0f4b5a6a