[roll] Roll fuchsia [ffx][package] Depend on test empty-repo
The coverage builder appears to be getting confused because of how we
were depending on the `empty-repo` artifacts from the `pkg` library. The
files are listed as a test dependency, but the `test_util` module is
exported as a non-test module. This tries to fix the test by directly
depending on these artifacts.

Original-Bug: 105539

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/707027
Original-Revision: 632f8b77a1ef380b798e0240a95193c3766ac0b3
GitOrigin-RevId: 10ff9262eb37108b111b5e42ce6892ca53c57787
Change-Id: I807fe03cbb4b111aba530721f35e6962268e18d4