[roll] Roll fuchsia [kernel][iommu] Fix discontiguous/contiguous logic.
Previously we would attempt to see if the range we were mapping was
contiguous. If it was, we would call MapDiscontiguous(). If it wasn't, we
would call MapContiguous().

Unsurprisingly, this worked OK if the mapping _was_ contiguous, but if
it wasn't, it would fail.

Original-Bug: 91878
Test: boot Atlas with IOMMU enabled, observe SDHCI driver works.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/706885
Original-Revision: c548849c299b242512683e804b69a5da6bc31318
GitOrigin-RevId: d89d5db60eab57b98c6b77e3e160138a9e85014b
Change-Id: I67f832e7f64e136e31f4d3a670fb0cfbff3f0a93