[roll] Roll fuchsia [settings] Fix flakiness in factory_reset_test
The factory reset FIDL does not return a result on the set FIDL call, so
clients are not guaranteed to know when their request has been
processed. This CL changes the mock to push the set calls it has
received to a channel so that the test can verify that the request was
processed and verify that the set call received the correct value.

Multiply: factory_reset_integration_test
Original-Fixed: 105502
Test: fx test factory_reset_integration_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/706902
Original-Revision: 2f3093c099a9b45ac1d049aebf3d55f3993d631f
GitOrigin-RevId: 2860adaee17f11d79d407270b2d001bb21eb63c3
Change-Id: Iaf851dfaf7a38e57e658e080f07b71e07065ff6b