[roll] Roll fuchsia [kernel] Simplify interrupt handlers
We can remove interrupt_eoi, as it was only needed by the hypervisor for
`zx_interrupt_bind_vcpu`. This allows us to simplify all of the
interrupt handlers in the kernel, and the interrupt handling logic in
the GIC code.

Original-Bug: 104712
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/706386
Original-Revision: 22c26ac955bd7114c75aa43d99b5d8dc99c55e05
GitOrigin-RevId: e60bc9129a14bad6d15b59ac66a60270deee4f07
Change-Id: Ice12c6a0e83303331a5dc93789d7d5c67821256e