[roll] Roll fuchsia [fidlgen] Leverage the `Declaration` interface (and friends)
This change introduces two successive extensions to the `Declaration`
that represent layouts and "resourceable" layouts. We this in hand, we
consolidate all "loop over each declaration in Root" logic to a single `Root` method that passes each `Declaration` - on which the caller can use type assertions to derive the concrete declaration type or the other two interfaces -  and then use this to substantially tidy up the sites that rely on such iterations.

This type will be of further value in other contexts where we'd want to
deal with the totality of declarations with minimal boilerplate (e.g.,
computing the dependency graph).

Original-Bug: 7660, 51002
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/704243
Original-Revision: 396626975e382104dbb320dd9540c4cdbb1c0611
GitOrigin-RevId: e4ecbf047d03dae7ad8b676e363a6bc816aa8a9b
Change-Id: Ib59779f86edf28a594734805dad08ce42e6acc6f