[roll] Roll fuchsia [virtualization][vsock] Wire in config provided listeners
This change wires config provided listeners in for the legacy and new
vsock device.

For the legacy device, config provided listeners are forwarded to the
host vsock. For the new device, config provided listeners are passed to
the VMM and onwards to the device.

Original-Bug: 97355
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705405
Original-Revision: f1f4662efd27b599b17ef03e5831f079a7a208fc
GitOrigin-RevId: 9fcfed1981bd14486e7ec1107bbbeb30734f8714
Change-Id: I6ba7f3fdd32f6ac614380088a2cbad459526d0fa