[roll] Roll fuchsia [mdns] make probe conflict checks case insensitive
This makes it so that when probe denials are sent with different case
(Example: host.local and hOsT.local) a device would recognize this as a
conflict/probe denial and choose a new hostname to probe with. Extended
current conflict tests to handle this scenario.

Original-Bug: 105386
Test: Manually issued case insensitive probe denials and confirmed
device recognized the denial and sent a new probe with a new hostname

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705105
Original-Revision: 0e8cefc2ae572432e7f92f5e2276b9d32d8ae807
GitOrigin-RevId: 1e1ae5e1353b5543ddb07f528314b6606b429710
Change-Id: Ic9180eea8dfad68b17eb4793d6f6b07e4134a067