[roll] Roll fuchsia [universe-resolver] Simplify .cmls
The base-only .cmls are no longer used, because the universe-resolver
no longer has a base-only mode, it always uses
fuchsia.pkg.PackageResolver (which in practice always comes from the
pkg-resolver) for package resolution.

The incoming fuchsia.component.resolution.Resolver capability is no
longer used (it used to be the Resolver protocol capability from the
base-resolver when universe-resolver had a base-only mode).

The for-test .cml was the same as the regular .cml.

example.cml was referencing a full-resolver component which does not
exist in the package.

Original-Bug: 105528
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705684
Original-Revision: 7b12c588f52c8df9134cde654084a877181119c2
GitOrigin-RevId: 1fc819ff3974aa2c14d53689891afe0902399081
Change-Id: I85c65e38bda5f96a7ed92efc35e19d240234c5a7