[roll] Roll fuchsia [netstack-gotests] Avoid deadlock on waiter events
TestTCPEndpointMapConnect is hitting a deadlock because the test is
written to assume at most 2 events are sent to the waiter constructed
for the test, but the waiter could be notified up to 3 times. The fix
is to increase the buffer size of the channel the events are being
written to.

Original-Fixed: 97765
Multiply: netstack-gotests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/706931
Original-Revision: c9112cddf9bf8291600f32f22f011f977735e2f9
GitOrigin-RevId: a418e9a79c35bffaa32481f5925463d127cf474c
Change-Id: Ib62a5f085371090f853da0d85c011e27e9d37616