[roll] Roll fuchsia [test_manager] Add scheduling options method to test_manager FIDL
This change adds a new method to the RunBuilder protocol called
WithSchedulingOptions, which is intended to provide test_manager
clients with an interface to specify whether or not they want
parallel test suite execution. This method receives as input a
struct containing a single field of type SchedulingOptions.
SchedulingOptions is a table that specifies whether maximum number of
test suites to execute in parallel. If this value is 0, the choice is
made by the server side. This change to test_manager's FIDL requires
us to handle the WithSchedulingOptions method in test manager. To
handle this, this CL adds a WithSchedulingOptions branch in
run_test_manager that invokes the unimplemented macro. Adding this
method also required some changes in the run_test_suite library to
satisfy Rust's exhaustive enum rule.
This change is part of a series of changes to properly route scheduling
options from clients to test_manager and ensuring that test_manager
behaves accordingly to clients' requests. This change only sets up the
FIDL method required to do this. Future work involves modifying the
run_test_suite library shared by the run_test_suite binary and ffx test
to call WithSchedulingOptions and modifying test_manager to use the
requested scheduling strategy. Please refer to the following desgin doc
for more information: go/tq-test-scheduler-prototype

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/707022
Original-Revision: 55e4c26f2936f5fa377caa02eac06999b1d36140
GitOrigin-RevId: 8c22559ec347bee73f6680839486818a3a0da42b
Change-Id: Ifa04e40c2b324ecf89553dba47b990a3316b744c