[roll] Roll fuchsia [assembly] Remove image assembly comparison safety net
The compare_image_assembly_configs() safety net has been growing less
useful, as more exceptions are needing to be added to it.  Moving
forward, as packages are moved from the legacy bundle to the common
platform bundles, and as more structured configuration is added to the
system, this is a less and less useful safety net, and a before vs.
after comparison is needed instead (but cannot be expressed in a
single build run, and so must be performed outside of the build
system).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705582
Original-Revision: d7ea08575548e7e82033b406b7fd72eea64b2ded
GitOrigin-RevId: a1e52392a4bf980039b6617f1750e9f0641e2bee
Change-Id: Ia39e3c1385bea5c9cbc6cdeee2d98a04e5bad6ee