[roll] Roll fuchsia [scrutiny][verifiers] Remove /core/universe-resolver allowlist
Now that full-resolver-env uses /core/universe-resolver to resolve
fuchsia-pkg URLs (because /core/full-resolver was deleted), this
instance of the allowlist has largely become a change detector for
children of /core. The other instances of this allowlist are not
changed and still restrict the use of /core/universe-resolver.

Original-Bug: 105528
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/707304
Original-Revision: d6e6a417ce1fc579673f1b340ac5eea9700696da
GitOrigin-RevId: c94df25d9773e8e07f278677b77fb366a889cd2e
Change-Id: Ic8d6687b03a18bc54c5902338eb96f1d1c8b23c5