[roll] Roll fuchsia [pci] Change policy on when we disable devices
It's possible that a device is enabled by the bootloader, or in use by
the kernel. If we probe the device and are unable to allocate an
existing window for it then we will now just leave it alone and no
longer expose any bind points for drivers. To enable this some changes
were made to cut down on the number of command register checks that were
needed, and an RAII Io disabler has been introduced to ensure the state
before and after our probing is coherent. Further improvements have been
made to cut down on command register modifications at the same time.

This should resolve issues on systems where the kernel carves out space
for something like a UART and the PCI driver is not able to maintain
that allocation with its own bookkeeping.

Original-Fixed: 96645
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/679687
Original-Revision: fe23d8f66076eedd168139779a106dc92540a0eb
GitOrigin-RevId: 4356a33a3a499096d35fe0781fed1a6d53d3209f
Change-Id: I15a8f2ac463939e1a8756476dfa5410147b0805f