[roll] Roll fuchsia [build] Allow specifying image name
The image name is use as a selector in a metadata file. Before,
the name was presumed to be "kernel" which is correct almost
always.

In some cases, we use a modified zbi image, which has a custom image
name.

If the image name is not set, will default to he label name. Which
matches zbi_tests at this moment. The default for kernel image,
is "kernel" hardcoded.

Test: fx build
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705262
Original-Revision: 806713e4620073e020d7273650b702e8237d10ce
GitOrigin-RevId: fc56b6053f9c170bca50a444dfa4dcd47657042a
Change-Id: I7200173da43f91cc206979a0f0dde8912d37dbc7