[roll] Roll fuchsia [virtualization] Allow some gpu/input queues to be optional.
The zircon drivers don't configure all the queues. With the virtio-gpu
we don't configure the cursorq and with the input device we don't
configure the statusq. The `take_stream` method will return `None` for
these queue when running against a zircon guest so we need to teach the
rest of the device to handle this situation.

Original-Bug: 104229, 102870
Test: fx test device_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705532
Original-Revision: d81e1876e1ec61ecbf262c51a46908715516cfb3
GitOrigin-RevId: 84373d9d8b71ac9e3b8f2c12c5963a036ea215d9
Change-Id: I1e62c7086d6ea333cdc841a1d431ac1e87bc94d5