[roll] Roll fuchsia [blobfs][host] Parallelize chunked compression
Add a new chunked compressor that uses a thread pool to compress the
chunks of the chunked compression format in parallel.

The blobfs host tool already compresses blobs in parallel but each blob
was being processed in a single thread. For builds with a few large
blobs, the entire build would be waiting for a couple threads to
compress these large blobs. The MultithreadedChunkedCompressor allows
the blobfs host tool to use all of threads on the system to compress
these large blobs which significantly reduces the build time of blobfs
images.

Original-Bug: 105094
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705682
Original-Revision: 25f4ccef11e90037c104d8f5a5f355f9c347287f
GitOrigin-RevId: 79a60546581d43024e28e3cd6817080ef7513f49
Change-Id: I19dccf47898ec7c96ae56659d2857877f19deb3a