[roll] Roll fuchsia [component][fuzzing] Add fuzz-test-runner
This CL adds a fuzz-test-runner as part of the component testing
framework. This runner acts almost identically to the elf-test-runner,
except that it additionally creates a channel, installs one end in the
started fuzz test, and passes the other to the fuzz-manager for use with
the fuchsia.fuzzer.Controller protocol.

Incidentally, it silences a warning from the transceiver if the
Controller client decides it doesn't care about receiving data it
requested and drops the connection (as happens in the fuzz-test-runner
tests).

Original-Bug: 24740
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/558392
Original-Revision: 34f4c34c57001595b1b45ab50cd48659f1fdde7a
GitOrigin-RevId: c7f2f02ac74edc174f9d5bfeb3f146654e9f9244
Change-Id: Ia13e282371af5ac3966299f74ceb1426b7612732