[roll] Roll fuchsia [ffx][isolate] Allow overriding ascendd socket
The isolate library configuring the socket through the environment
variable prevents users of it from overriding the socket path, as is
done in the daemon self-test, causing it to not work correctly (though
apparently only some of the time).

Setting it both ways is no longer necessary, ASCENDD and overnet.socket
have a priority ordering and it is sufficient to set it as a default
configuration that can be overridden with command line arguments now.

Original-Bug: 104562

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/708102
Original-Revision: d9a5c73fa1c20bac753dc7b79ffbdd9d19f49156
GitOrigin-RevId: 5b81f42f1fd637d31b123e3164367cd0638fb3b9
Change-Id: Ie85c9bb871b3e463482bbc5b18f23ede0232efe8