[roll] Roll fuchsia [build] In-tree C++ builds at "HEAD" API level.
In order for C++ libraries in the SDK to make use of pending FIDL APIs
when built in-tree, it must be possible for the C++ API to mark its
availability to match.

clang does not have a concept of a HEAD API level, so here we emulate
it using a very large integer.

Original-Bug: 104513
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/708010
Original-Revision: 3dd9e744c5889de2681c18cae5b3d616e6a97417
GitOrigin-RevId: 4d3eeb094c795ce08769bd1e7505a3f67b844495
Change-Id: Ifb9ae53a2f34c1f2b402eee3c208c1c533955e27