[roll] Roll fuchsia [bt][profiles] Migrate bt-a2dp-looback to CFv2
Define v2 manifests for the loopback test and A2DP sink/source
components.

Migrate the loopback shell package to a Fuchsia test. This test can then
be invoked by SL4F to start streaming.

Remove the optional argument to specify a streaming duration. This is a
fixed value specified in the E2E test code and is never used otherwise.

Update test_manager to provide the SessionAudioConsumerFactory
capability to system tests. This allows the A2DP sink component to
stream audio.
Add a TODO pointing to the mediasession migration so that this
capability can be offered hermetically once the component is migrated to
CFv2.

Don't include the test by default as it requires hardware to pass.

Original-Bug: 74812
Test: fx test bt-a2dp-loopback-test
Test: Manual, fx test bt-a2dp-loopback-test on sherlock and verify test
streams for 10seconds with successful exit.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/706486
Original-Revision: f1c0ef9d2b0379250bd152316662787a8f3004f0
GitOrigin-RevId: 65b8338f12d789138d92586c269e7d4804856c4a
Change-Id: Ib797db05e75de6b846e0c4e21838d68d55f996f1