[roll] Roll fuchsia [ffx][config] Clean up edge-case lockfile failures
If we fail to write out details to the lockfile (which should be pretty
rare if we managed to open the file to begin with, but seems like it
might be happening sometimes), we should remove the incorrect lockfile
before returning Err.

We should also flush the file to disk immediately and not return until
the file is flushed.

This hopefully increases the reliability of the lockfiles, and produces
useful output more of the time when something goes wrong. To help with
that too, let's output the current process' pid in the error as well so
it's a little more visible if a process is deadlocking itself.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/708533
Original-Revision: 3908df51c12439e301302eea7b3c76744e3ee11a
GitOrigin-RevId: 7f27461ac3946310c416def9d503504517046f05
Change-Id: Iee84668dfe36cdbd95f1efbd167f69de9536c124