[roll] Roll fuchsia [virtualization] Enable the rust virtio-gpu/input devices by default
I'm in the process of enabling the integration tests for these
components. Since the older c++ ones had many tests disabled and they
rely on different APIs (gfx vs flatland), it will be simpler to enable
these devices along with the tests.

The risk should be limited since these devices do not see heavy use
(they're unused in workstation) and have good component level testing.
Compared to the c++ devices, we do not yet support mouse or cursor
commands but these should not be a large amount of work to re-implement
on top of flatland.

Original-Bug: 102870, 104229
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/706914
Original-Revision: e95712f541689802d2019581041c3903784e16f3
GitOrigin-RevId: 9c785fff830271d68dde95dc5951fb36e513b713
Change-Id: I82439b1a22df98cb9e40165c67ac65f8bf3e08cc