[roll] Roll fuchsia [zbi] Remove dependencies on other zircon headers
The ZBI spec should be a leaf, dependency-wise, and not depend on other
Zircon types and macros. In particular, this makes it difficult to copy
the ZBI-related headers on their own to bootloader codebases (today the
whole 'sysroot' is defensivly taken along with it).

In the not-too-distant future we will generate <zircon/boot/*> (from
FIDL), at which point we will have better tooling at our disposal to
ensure that these files do not pick up new dependencies.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/709265
Original-Revision: 0628d21996dcbc0caaf3fd689059e6e38f1c6eb7
GitOrigin-RevId: 85a7e44bf9bc7f88e9b9199f91f831a28bd1b5bb
Change-Id: Idf7c19f4ef05e13bde193707a7320ed58eb40ace